% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_class.R
\name{add_class}
\alias{add_class}
\title{Add A New Class}
\usage{
add_class(x, define_class)
}
\arguments{
\item{x}{An object.}

\item{define_class}{A new self-defined class}
}
\description{
Add A New Class
}
\examples{
a <- matrix(rnorm(1000), nrow = 100)
a <- add_class(a, 'my_matrix')
}
