% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_utils.R
\name{add_data_to_model}
\alias{add_data_to_model}
\title{Add Gene Expression Data to Model}
\usage{
add_data_to_model(model, dataset)
}
\arguments{
\item{model}{A dynwrap::with_trajectory object generated by \code{\link[dynwrap]{infer_trajectory}}.
function.}

\item{dataset}{A dynwrap::data_wrapper object generated by \code{\link[dynwrap]{infer_trajectory}}.
function.}
}
\value{
A new dynwrap::with_trajectory object.
}
\description{
This function is used to add the gene expression data into the result of
trajectory inference generated by \code{\link[dynwrap]{infer_trajectory}} function.
}
\examples{
# Generate a reference data
set.seed(1)
a <- matrix(rpois(n = 2500, lambda = 2), nrow = 50)
rownames(a) <- paste0("cell", 1:ncol(a))
colnames(a) <- paste0("gene", 1:nrow(a))
dataset_ref <- dynwrap::wrap_expression(
  counts = a,
  expression = log2(a+1)
)
# Trajectory inference
}
