% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_dependent_effect}
\alias{add_dependent_effect}
\title{Add Depended Effect to Target}
\usage{
add_dependent_effect(data, effect, options, config)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{effect}{vector with integer effects}

\item{options}{list with user-defined command line arguments}

\item{config}{list with further arguments}
}
\description{
In scenario 2, additionally to the independent effects, there exist also
dependent effects. This function adds these dependent effects to the target
variable: At a specific group and at specific points of time, integer
effects are added to the target variable. These effects depend on the ones
provided as argument to this function.
}
\details{
Currently the dependent effect is simply half of the given effect. It is
ensured that the depended effect consists only of integers by applying
integer division.

config$dependent_effect_time is a vector that specifies the points of time
at which effects are added.
config$time_variable is the name of the column in data that contains times.

config$placebo_group is the name of the group to which effects are added.
config$group_variable is the name of the column in data that contains groups.

Note that data will be modified in place.
}
