% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_main_effect}
\alias{add_main_effect}
\title{Add Independent Random Effect to Target}
\usage{
add_main_effect(data, params, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{params}{named vector that maps parameter names to parameter values}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\value{
vector with the effects that were added to the target variable
}
\description{
At a specific group and at specific points of time, add random effects
to the target variable.
}
\details{
\code{options$effect_type} is the name of the distribution of the added effects.
The params argument provides additional parameters for this distribution.
\code{options$target} is the user-defined name of the target variable in \code{data}
to which effects are applied.

\code{config$main_effect_time} is a vector that specifies the points of time at
which effects are added.
\code{config$time_variable} is the name of the column in \code{data} that contains
times.
\code{config$placebo_group} is the name of the group to which effects are added.
\code{config$group_variable} is the name of the column in \code{data} that contains
groups.

Note that \code{data} will be modified in place.
}
