% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_s2_effect}
\alias{add_s2_effect}
\title{Add Depended Effect to Target}
\usage{
add_s2_effect(data, effect, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{effect}{vector with random effects}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\description{
In scenario 2, additionally to the independent effects, there exist also
dependent effects. This function adds these dependent effects to the target
variable: At a specific group and at specific points of time, random
effects are added to the target variable. These effects depend on the ones
provided as argument to this function. The dependend effect is simply half
of the given effect.
}
\details{
\code{config$s2_effect_time} is a vector that specifies the points of time
at which effects are added.
\code{config$time_variable} is the name of the column in \code{data} that contains
timepoints.
\code{config$placebo_group} is the name of the group to which effects are added.
\code{config$group_variable} is the name of the column in \code{data} that contains
groups.

Note that \code{data} will be modified in place.
}
