% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_s3_effect}
\alias{add_s3_effect}
\title{Add Depended Effects to Target for Different Timepoints}
\usage{
add_s3_effect(data, effect, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{effect}{vector with random effects}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\description{
In scenario 3, additionally to the independent effects, there exist also
dependent effects. This function adds these dependent effects to the target
variable: At a specific group and at specific points of time, random
effects are added to the target variable. These effects depend on the ones
provided as argument to this function. Note that, in contrast to scenario
2, the effects added by this function consist of integers only.
}
\details{
The dependend effect for the second timepoint is half of the given effect.
The dependend effect for the fourth timepoint is half of the given effect
plus a rounded standard normally distributed random effect.

\code{config$s3_effect_time_a} is a vector that specifies the points of time
at which effects are added.
\code{config$s3_effect_time_b} is a vector that specifies the points of time
at which effects are added.
\code{config$time_variable} is the name of the column in \code{data} that contains
timepoints.
\code{config$placebo_group} is the name of the group to which effects are added.
\code{config$group_variable} is the name of the column in \code{data} that contains
groups.

Note that \code{data} will be modified in place.
}
