% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_definition.R
\name{authors_definition}
\alias{authors_definition}
\title{Authors information}
\usage{
authors_definition(first, last, email = NULL, github = NULL, orcid = NULL)
}
\arguments{
\item{first}{First name.}

\item{last}{Last name.}

\item{email}{Email address.}

\item{github}{Github URL.}

\item{orcid}{Orcid URL.}
}
\value{
A list.
}
\description{
Authors information
}
