% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binarize_target}
\alias{binarize_target}
\title{Dichotomize Target Variable}
\usage{
binarize_target(data, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\description{
If desired, replace the target variable values with 0/1 values.
}
\details{
Target variable values are supposed to be divided into blocks of equal size.
Each block combines measurements for one subject at different points of
time. Whether a measurement is replaced by 0 or by 1 depends on its
magnitude relative to the first value in the same block (i.e., relative to
the baseline measurement). After all target values have been replaced by
either 0 or 1, the baseline measurement is removed from data. This is
because all baseline value would equal 0 (as they cannot decrease relative
to themselves), and hence, do not contain additional information.

\code{options$target} contains the name of the target variable.
\code{options$binarize} determines if the target should be binarized or not
(\code{TRUE}/\code{FALSE}).
\code{config$binary_threshold} a relative measure that determines whether a
target value is replaced by 0 or 1.
\code{config$blocklength} is the number of measurements in a block that refer to
one subject.
\code{config$baseline_time} contains the points in time at which baseline
measures were collected.

Note that \code{data} is modified in place.
}
