% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_batch_properties.R
\name{calculate_batch_properties}
\alias{calculate_batch_properties}
\title{Calculate Batch Quantifying Metrics}
\usage{
calculate_batch_properties(data, batch_info, k, verbose = TRUE)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{batch_info}{Batch assignment of every cells in columns of matrix.}

\item{k}{k nearest neighborhoods of every cell.}

\item{verbose}{Whether messages are returned during the process.}
}
\description{
Calculate Batch Quantifying Metrics
}
