% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_properties.R
\name{calculate_CDI}
\alias{calculate_CDI}
\title{Calculate Clustering Deviation Index (CDI)}
\usage{
calculate_CDI(data, cluster_info, batch_info = NULL)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{cluster_info}{Cluster(or group) assignment of every cells in columns of matrix.}

\item{batch_info}{Default is NULL. Batch information of every cells.}
}
\value{
A dataframe of CDI values
}
\description{
Calculate Clustering Deviation Index (CDI)
}
\references{
Fang J, Chan C, Owzar K, et al. Clustering Deviation Index (CDI): A robust and accurate unsupervised measure for evaluating scRNA-seq data clustering. bioRxiv, 2022. \url{https://doi.org/10.1101/2022.01.03.474840}
Github URL: \url{https://github.com/jichunxie/CDI}
}
