% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_properties.R
\name{calculate_CH_index}
\alias{calculate_CH_index}
\title{Calculate Calinski-Harabasz Index}
\usage{
calculate_CH_index(data, cluster_info)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{cluster_info}{Cluster(or group) assignment of every cells in columns of matrix.}
}
\description{
Calculate Calinski-Harabasz Index
}
