% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_properties.R
\name{calculate_cluster_properties}
\alias{calculate_cluster_properties}
\title{Summarize the Ability of Simulating Clusters}
\usage{
calculate_cluster_properties(
  data,
  dist = NULL,
  cluster_info,
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{A matrix or dataframe of gene expression.}

\item{dist}{Optionally, if NULL, the distance matrix is computed.}

\item{cluster_info}{The vector of characters which each cell belongs to.}

\item{threads}{How many cores used for parallel computation.}

\item{verbose}{Whether the messages of execution process are returned.}
}
\value{
A list of metric results
}
\description{
Summarize the Ability of Simulating Clusters
}
