% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_properties.R
\name{calculate_connectivity}
\alias{calculate_connectivity}
\title{Calculate Connectivity}
\usage{
calculate_connectivity(data, cluster_info)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{cluster_info}{Cluster(or group) assignment of every cells in columns of matrix.}
}
\description{
Calculate Connectivity
}
