% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_properties.R
\name{calculate_ROGUE}
\alias{calculate_ROGUE}
\title{Calculate ROUGE Value
Calculate entropy-based metric for assessing the purity of single cell populations.}
\usage{
calculate_ROGUE(data, cluster_info)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{cluster_info}{Cluster(or group) assignment of every cells in columns of matrix.}
}
\value{
A numeric ranged from 0 to 1.
}
\description{
Calculate ROUGE Value
Calculate entropy-based metric for assessing the purity of single cell populations.
}
\references{
Liu B, Li C, Li Z, et al. An entropy-based metric for assessing the purity of single cell populations. Nature communications, 2020, 11(1): 1-13. \url{https://doi.org/10.1038/s41467-020-16904-3}
Github URL: \url{https://github.com/PaulingLiu/ROGUE}
}
