% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cluster_properties.R
\name{calculate_silhouette}
\alias{calculate_silhouette}
\title{Calculate Average Silouette Width}
\usage{
calculate_silhouette(data, cluster_info)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{cluster_info}{Cluster(or group) assignment of every cells in columns of matrix.}
}
\description{
Calculate Average Silouette Width
}
