% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_trajectory_properties.R
\name{calculate_trajectory_properties}
\alias{calculate_trajectory_properties}
\title{Calculate Metrics for Comparison of Two Trajectories}
\usage{
calculate_trajectory_properties(
  ref_data,
  ref_data_grouping = NULL,
  sim_data,
  sim_data_grouping = NULL,
  algorithm = "Hungarian",
  seed = 1,
  verbose = TRUE
)
}
\arguments{
\item{ref_data}{A gene matrix or the trajectory object generated by \code{dynwrap::wrap_expression} from real data}

\item{ref_data_grouping}{The labels of cells in real data}

\item{sim_data}{A gene matrix or the trajectory object generated by \code{dynwrap::wrap_expression} from simulated data}

\item{sim_data_grouping}{The labels of cells in simulated data}

\item{algorithm}{Optional. Which algorithm used for matching cells in simulated and real data. Improved_Hungarian (default), Hungarian.}

\item{seed}{Random seed}

\item{verbose}{Whether the messages are returned to users when processing}
}
\value{
A list
}
\description{
Calculate Metrics for Comparison of Two Trajectories
}
