% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_data_properties.R
\name{cell_properties}
\alias{cell_properties}
\title{Calculate Cell Properties}
\usage{
cell_properties(data, verbose = FALSE)
}
\arguments{
\item{data}{A count matrix.}

\item{verbose}{Whether messages are returned during the process.}
}
\value{
A list
}
\description{
Calculate Cell Properties
}
\examples{
data <- matrix(rpois(100*100, 2),
               nrow = 100,
               dimnames = list(paste0('Gene', 1:100),
                               paste0('Cell', 1:100)))
result <- cell_properties(data, verbose = TRUE)
}
