% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_parameters.R
\name{change_parameters}
\alias{change_parameters}
\title{Change parameters}
\usage{
change_parameters(function_expr, other_prior, step)
}
\arguments{
\item{function_expr}{The function expression used for get the parameters.}

\item{other_prior}{The list of new parameters remained to replace.}

\item{step}{Which step you are in.}
}
\description{
Change parameters
}
