% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_parameters.R
\name{change_scGAN_parameters}
\alias{change_scGAN_parameters}
\title{Change Parameters About scGAN}
\usage{
change_scGAN_parameters(project_name, ...)
}
\arguments{
\item{project_name}{Name your experiment}

\item{...}{Other parameters}
}
\description{
Change Parameters About scGAN
}
