% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_prior_info.R
\name{check_prior_info}
\alias{check_prior_info}
\title{Check Prior Information}
\usage{
check_prior_info(method, step, other_prior)
}
\arguments{
\item{method}{Which method}

\item{step}{Simulation or estimation}

\item{other_prior}{Prior information}
}
\description{
Check Prior Information
}
