% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.r
\name{ci.default}
\alias{ci.default}
\title{Confidence Intervals Using z-Score or t-Score}
\usage{
\method{ci}{default}(x, alpha = 0.05, z_score = TRUE)
}
\arguments{
\item{x}{An object of class `robust`.}

\item{alpha}{The significance level (default is 0.05).}

\item{z_score}{Logical indicating whether to use z-scores (default is TRUE). If FALSE, t-scores are used.}
}
\value{
A matrix of confidence intervals with lower and upper bounds.
}
\description{
This method calculates confidence intervals for a robust object, using either z-scores or t-scores.
}
