% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.r
\name{ci}
\alias{ci}
\title{Generic Function for Confidence Intervals}
\usage{
ci(x, alpha = 0.05, z_score = TRUE)
}
\arguments{
\item{x}{An object for which confidence intervals are to be calculated.}

\item{alpha}{The significance level (default is 0.05).}

\item{z_score}{Logical indicating whether to use z-scores (default is TRUE). If FALSE, t-scores are used.}
}
\value{
A matrix of confidence intervals, depending on the object type.
}
\description{
This is a generic function to calculate confidence intervals. It dispatches methods based on the class of the input.
}
