% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simutils.R
\name{clear_sim_dir}
\alias{clear_sim_dir}
\title{Clear simulation directory}
\usage{
clear_sim_dir(sim_dir, debug = 1)
}
\arguments{
\item{sim_dir}{A character string representing the directory path of the simulation.}

\item{debug}{An integer for debugging level. Defaults to 1.}
}
\value{
Returns (invisibly) the output of \code{file.remove()}.
}
\description{
Clears all temporary files (indicating in progress) from a simulation directory.
}
