% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.robust.r
\name{coef.robust}
\alias{coef.robust}
\title{Extract Coefficients from a Robust Object}
\usage{
\method{coef}{robust}(object, ...)
}
\arguments{
\item{object}{An object of class `robust` containing the model's robust results.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A vector of coefficient estimates.
}
\description{
This function extracts the coefficient estimates from a robust object created using the `robust` function.
}
