% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_power}
\alias{compute_power}
\title{Simulation-based Computation of Power}
\usage{
compute_power(data, params, options, config)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{params}{named vector that maps parameter names to values}

\item{options}{list with user-defined command line arguments}

\item{config}{list with further arguments}
}
\value{
vector with average power values for both periods
}
\description{
For a given number of repetitions, firstly permute the target variable and
secondly, add effects to selected values such that a a situation in which
H1 holds is established. Perform hypotheses tests for both periods of the
trial separately. Store all test results and return the average for both
periods.
}
\details{
The additive random effect is specified by a combination of user input (the
attribute options$effect) and the params argument.

options$target contains the name of the target variable.
config$repetitions is the number of repetitions to perform (i.e., the
number of tests performed for each period)
config$blocklength is the number of measurements in a block that refer to
one subject.
config$alpha is the expected type I error rate.

Moreover, options and config must contain all attributes required by
add_effect.
}
