% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_rejection_rate}
\alias{compute_rejection_rate}
\title{Simulation-Based Computation of H0 Rejection Rate}
\usage{
compute_rejection_rate(data, params, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{params}{named vector that maps parameter names to values or \code{NULL}}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\value{
vector with average power values for both periods
}
\description{
For a given number of repetitions, firstly permute the target variable and
secondly, add effects to selected values such that a a situation in which
H1 holds is established. Perform hypotheses tests for both periods of the
trial separately. Store all test results and return the average for both
periods.
}
\details{
The additive random effect is specified by a combination of user input (the
attribute \code{options$effect}) and the \code{params} argument. If \code{NULL} is assigned
to the \code{params} argument, then no random effect will be added. This way, the
type-I error can be simulated.

Moreover, users can choose to binarize the target variable (cf.
\code{binarize_target}).

\code{options$target} contains the name of the target variable.
\code{config$repetitions} is the number of repetitions to perform (i.e., the
number of tests performed for each period)
\code{config$blocklength} is the number of measurements in a block that refer to
one subject.
\code{config$alpha} is the expected type I error rate.
\code{options$binarize} determines whether the target should be binarized (cf.
\code{binarize_target}).

Moreover, \code{options} and \code{config} must contain all attributes required by
\code{add_effect}, \code{binarize_target}, \code{discard_baseline}and \code{perform_test}.
}
