% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_voronoi_neighbours.R
\name{compute_voronoi_neighbours}
\alias{compute_voronoi_neighbours}
\title{Compute Voronoi neighbours of antenna sites.}
\usage{
compute_voronoi_neighbours(simData, PLOT = FALSE, latlong = FALSE, R = 6371)
}
\arguments{
\item{simData}{\code{list} with components map, network, coverage, grid,
individuals as output by function \link{read_simData}.}

\item{PLOT}{logical (default FALSE) to indicate whether to plot the
tesselation or not.}

\item{latlong}{logical (default FALSE) to indicate whether \code{x} and
\code{y} refer to longitude and latitude.}

\item{R}{numeric vector of lenght 1 denoting the Earth's radius in km.}
}
\value{
list with:
Adj - Adjacency matrix
NumN - number of neighbours per antenna
D - Distance matrix
}
\description{
Find neighbours of antennas and number of neighbours for each
antenna in addition to the distance matrix. If latlong is TRUE then it
returns the distance matrix in km.
}
\examples{
data <- data.frame(i = 1:6, x = c(10,2,1,3,6,8), y = c(1,10,5,7,4,8))
a <- compute_voronoi_neighbours(data, TRUE)


filename_map      <- c(
 xml= system.file("extdata/input_files", "map.xml", package = "simutils"),
 xsd= '')
  
filename_network  <- c(
 csv= system.file("extdata/output_files/antennas.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/antennas_dict.xml", 
                   package = "simutils"))
                   
filename_signal <- c(
 csv= system.file("extdata/output_files/SignalMeasure_MNO1.csv", 
                   package = "simutils"),
 xml= system.file("extdata/metadata/output_files/SignalMeasure_dict.xml", 
                   package = "simutils"))
                 
filename_coverage <- c(
 csv= system.file("extdata/output_files", "AntennaCells_MNO1.csv", 
                  package = "simutils"),
 xml= system.file("extdata/metadata/output_files/AntennaCells_dict.xml", 
                  package = "simutils"))
                       
filename_grid <- c(
  csv= system.file("extdata/output_files/grid.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/grid_dict.xml", 
                   package = "simutils")) 

filename_individ <- c(
  csv= system.file("extdata/output_files/persons_dash.csv", 
                   package = "simutils"),
  xml= system.file("extdata/metadata/output_files/persons_dash_dict.xml", 
                   package = "simutils"))   
                       
filenames <- list(
  map                = filename_map,
  network_parameters = filename_network,
  signal             = filename_signal,
  coverage_cells     = filename_coverage,
  grid               = filename_grid,
  individuals        = filename_individ)
  
simData <- read_simData(filenames, crs = 2062)
compute_voronoi_neighbours(simData, PLOT = TRUE)


}
