% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_voronoi_rings.R
\name{compute_voronoi_rings}
\alias{compute_voronoi_rings}
\title{Establish an integer-valued distance between Voronoi polygons.
This distance basically determines the Voronoi ring for each antenna.}
\usage{
compute_voronoi_rings(voronoi)
}
\arguments{
\item{voronoi}{\code{list} of \code{sf} objects:triangles.sf and polygons.sf with:
geometry - coordinates of triangles and polygons
info of antennas which compose the objects}
}
\value{
a list with the distance in rings between polygons.
}
\description{
Establish an integer-valued distance between Voronoi polygons.
This distance basically determines the Voronoi ring for each antenna.
}
\examples{
filename_map      <- c(
 xml= system.file("extdata/input_files", "map.xml", package = "simutils"),
 xsd= '')
  
filename_network  <- c(
 csv= system.file("extdata/output_files/antennas.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/antennas_dict.xml", package = "simutils"))
                   
filename_signal <- c(
 csv= system.file("extdata/output_files/SignalMeasure_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/SignalMeasure_dict.xml", package = "simutils"))
                 
filename_coverage <- c(
 csv= system.file("extdata/output_files", "AntennaCells_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/AntennaCells_dict.xml", package = "simutils"))

filename_events <- c(
 csv= system.file("extdata/output_files/AntennaInfo_MNO_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/events_dict.xml", package = "simutils"))
                       
filename_grid <- c(
  csv= system.file("extdata/output_files/grid.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/grid_dict.xml", package = "simutils")) 

filename_individ <- c(
  csv= system.file("extdata/output_files/persons.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/persons_dict.xml", package = "simutils"))   
                       
filenames <- list(
  map                = filename_map,
  network_parameters = filename_network,
  signal             = filename_signal,
  events             = filename_events,
  coverage_cells     = filename_coverage,
  grid               = filename_grid,
  individuals        = filename_individ)
  
simData <- simutils::read_simData(filenames, crs = 2062)
voronoi <- simutils::compute_voronoi_sf(simData)
compute_voronoi_rings(voronoi) -> x

}
