% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{data_conversion}
\alias{data_conversion}
\title{Convert Data Format}
\usage{
data_conversion(
  SCE_object,
  return_format = "SingleCellExperiment",
  in_docker = FALSE,
  local_path = NULL
)
}
\arguments{
\item{SCE_object}{A SingleCellExperiment object}

\item{return_format}{Which data format you want, list or Seurat or h5ad or
SingleCellExperiment}

\item{in_docker}{Logical. Do you process simulation step in Docker?}

\item{local_path}{If \code{in_docker} is TRUE, you must input the directory
path on your local device}
}
\value{
The object you want, list or Seurat or h5ad or SingleCellExperiment.
}
\description{
Convert Data Format
}
