% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_datasets.R
\name{default_datasets}
\alias{default_datasets}
\title{Default Datasets For Simulations}
\usage{
default_datasets(method)
}
\arguments{
\item{method}{Method name}
}
\description{
This function is used to get the default datasets for simulations.
}
\examples{
datasets <- default_datasets(method = "SCRIP")
}
