% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_parameters.R
\name{default_parameters}
\alias{default_parameters}
\title{Default Parameters of Simulation Methods}
\usage{
default_parameters(method)
}
\arguments{
\item{method}{Method name}
}
\description{
This function is used to get the default parameters of a simulation method.
}
\examples{
parameters <- default_parameters(method = "Splat")
}
