% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discard_baseline}
\alias{discard_baseline}
\title{Remove Baseline Measurements}
\usage{
discard_baseline(data, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\value{
the shrinked dataset (original dataset remains untouched)
}
\description{
After all target values have been replaced by either 0 or 1, the baseline
measurement can be removed from data. This is because all baseline values
equal 0 (as they cannot decrease relative to themselves), and hence, do not
contain additional information.
}
\details{
\code{options$binarize} determines if the target was be binarized or not (\code{TRUE}/
\code{FALSE}).
\code{config$time_variable} is the name of the variable containing timepoints in
the dataset.
\code{config$baseline_time} contains the points in time at which baseline measures
were collected.
}
