% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doCellFromXY}
\alias{doCellFromXY}
\title{Get cell number from x and y coordinates}
\usage{
doCellFromXY(ncols, nrows, xmin, xmax, ymin, ymax, x, y)
}
\arguments{
\item{ncols}{Number of colums of the raster grid.}

\item{nrows}{Number of rows of the raster grid.}

\item{xmin}{Minimal value of x coordinates.}

\item{xmax}{Maximal value of x coordinates.}

\item{ymin}{Minimal value of y coordinates.}

\item{ymax}{Maximal value of y coordinates.}

\item{x}{Vector of x coordinates.}

\item{y}{Vector of y coordinates.}
}
\description{
Get cell number from x and y coordinates
}
