% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{doXYFromCell}
\alias{doXYFromCell}
\title{Get x and y coordinates from cell ID number.}
\usage{
doXYFromCell(ncols, nrows, xmin, xmax, ymin, ymax, cell)
}
\arguments{
\item{ncols}{Number of colums of the raster grid.}

\item{nrows}{Number of rows of the raster grid.}

\item{xmin}{Minimal value of x coordinates.}

\item{xmax}{Maximal value of x coordinates.}

\item{ymin}{Minimal value of y coordinates.}

\item{ymax}{Maximal value of y coordinates.}

\item{cell}{Vector of cell ID numbers.}
}
\description{
Get x and y coordinates from cell ID number.
}
