% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exclude_na_blocks}
\alias{exclude_na_blocks}
\title{Remove Blocks Containing NA in Target}
\usage{
exclude_na_blocks(data, target, blocklength)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{target}{name of the target variable column in data}

\item{blocklength}{number of measurements in a block}
}
\value{
the input data reduced by excluded NA-blocks
}
\description{
If NAs occur in the target variable, the respective blocks will be removed
from the dataset (i.e., the NA-rows as well as their surrounding rows will
be removed)
}
