% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_deepest.R
\name{flatten_deepest}
\alias{flatten_deepest}
\title{Transform a list of lists whose final nodes are \code{list} into
a list of lists with this final nodes as different elements (e.g:
\code{character}).}
\usage{
flatten_deepest(list)
}
\arguments{
\item{list}{A \code{list} of lists which must be named.}
}
\description{
\code{flatten_deepest} receives a list of lists with final nodes
as \code{list} and transforms it into a list of lists with final nodes as
different elements, for example, as \code{character} class; keeping the
names and attributes' structure.
}
\details{
Return a \code{list} of \code{list}s with final nodes as different
classes (\code{character}, \code{integer}, ...) keeping the original names'
structure.
}
\examples{
rootPath  <- system.file(package = "simutils")  
xmlFile   <- file.path(rootPath, "extdata/input_files", "simulation.xml")
xmlObject <- as_list(xml2::read_xml(xmlFile))
flatten_deepest(xmlObject)
 
}
