% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_data_properties.R
\name{gene_properties}
\alias{gene_properties}
\title{Calculate Gene Properties}
\usage{
gene_properties(data, cpm_norm = TRUE, verbose = FALSE)
}
\arguments{
\item{data}{A count matrix.}

\item{cpm_norm}{Whether CPM normalization will be used. Default is TRUE.}

\item{verbose}{Whether messages are returned during the process.}
}
\value{
A list
}
\description{
Calculate Gene Properties
}
\examples{
data <- matrix(rpois(100*100, 2),
               nrow = 100,
               dimnames = list(paste0('Gene', 1:100),
                               paste0('Cell', 1:100)))
result <- gene_properties(data, verbose = TRUE)
str(result)
}
