% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_effect}
\alias{generate_effect}
\title{Generate Additive Random Effect}
\usage{
generate_effect(effect_type, n, params)
}
\arguments{
\item{effect_type}{name of the effect distribution}

\item{n}{size of the returned random sample}

\item{params}{named vector that maps parameter names to parameter values}
}
\value{
vector with a random sample of size n
}
\description{
The generated effects are supposed to be added to the target variable
afterwards.
}
\details{
The returned vector is rounded to one floating point digit.
}
