% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_manipulation.R
\name{get_default_value}
\alias{get_default_value}
\title{Get the Default Value}
\usage{
get_default_value(x)
}
\arguments{
\item{x}{An object created by \code{\link[simmethods:get_method]{simmethods::get_method()}}}
}
\value{
A list
}
\description{
Get the Default Value
}
\examples{
# x <- simmethods::get_method()
# params <- get_default_value(x)
}
