% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_manipulation.R
\name{get_method_definition}
\alias{get_method_definition}
\title{Get Information of Methods}
\usage{
get_method_definition(method = "all")
}
\arguments{
\item{method}{One or more method names. Default is all.}
}
\value{
A list.
}
\description{
Get Information of Methods
}
\examples{
get_method <- get_method_definition(method = c("splat", "PROSSTT"))
}
