% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mnd.R
\name{get_mnd}
\alias{get_mnd}
\title{Get mobile network data.}
\usage{
get_mnd(simData, devices, t_range, groundTruth = FALSE)
}
\arguments{
\item{simData}{list of different information elements from the simulation (see
\ref{\code{read_simData}}).}

\item{devices}{character vector with the codes of the devices for which the data are to be
obtained; if nothing is specified, all are taken.}

\item{t_range}{numeric vector with the range of time instants whose events are to be obtained;
it has length 2, where first component contains the initial instant and the second the final one.
By default, the whole range of times is chosen.}

\item{groundTruth}{logical vector indicating if the ground truth is included in the output or not}
}
\description{
Create a named list of two elements related with the simulation input. The first element is
related to the observed mobile network data (mnd) whereas the second component contains the
ground truth. It basically contains the events detected in the network and the ground truth that
is fill by events characteristics matched with PersonId
}
\examples{
filename_map      <- c(
 xml= system.file("extdata/input_files", "map.xml", package = "simutils"),
 xsd= '')
  
filename_network  <- c(
 csv= system.file("extdata/output_files/antennas.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/antennas_dict.xml", package = "simutils"))
                   
filename_signal <- c(
 csv= system.file("extdata/output_files/SignalMeasure_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/SignalMeasure_dict.xml", package = "simutils"))
                 
filename_coverage <- c(
 csv= system.file("extdata/output_files", "AntennaCells_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/AntennaCells_dict.xml", package = "simutils"))

filename_events <- c(
 csv= system.file("extdata/output_files/AntennaInfo_MNO_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/events_dict.xml", package = "simutils"))
                       
filename_grid <- c(
  csv= system.file("extdata/output_files/grid.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/grid_dict.xml", package = "simutils")) 

filename_individ <- c(
  csv= system.file("extdata/output_files/persons.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/persons_dict.xml", package = "simutils"))   
                       
filenames <- list(
  map                = filename_map,
  network_parameters = filename_network,
  signal             = filename_signal,
  events             = filename_events,
  coverage_cells     = filename_coverage,
  grid               = filename_grid,
  individuals        = filename_individ)
  
simData <- simutils::read_simData(filenames, crs = 2062)

get_mnd(simData)

get_mnd(simData, devices = c('241', '254'), t_range = c(56,65), groundTruth = TRUE)

}
