% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{get_period_frame}
\alias{get_period_frame}
\title{Summarise Trial Period in a Dataframe}
\usage{
get_period_frame(data, config)
}
\arguments{
\item{data}{\code{data.table} with the study data of one trial period}

\item{config}{global config object

\code{config$subject_variable} is the name of the variable that identifies
subjects in data.
\code{config$group_variable} is the name of the group variable in \code{data}.}
}
\value{
(printable) data.frame with the desired information.
}
\description{
Create a data.frame with information about \code{data}, which is supposed to
contain only records from a single trial period (i.e., period splitting
must be performed in prior to using this function). The resulting data.frame
will have a row for each subject and three columns. Two of the columns are
named according to \code{config$subject_variable} and \code{config$group_variable},
respectively. The last one is named \code{Count} and contains a record count of
the respective subject.
}
