% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_period_info}
\alias{get_period_info}
\title{Summarise Dataset Containing a Single Trial Period}
\usage{
get_period_info(data, config)
}
\arguments{
\item{data}{data.table with the study data}

\item{config}{global config object

config$subject_variable is the name of the variable that identifies subjects
in data.
config$group_variable is the name of the group variable in data.}
}
\value{
printable string with dataset info.
}
\description{
Create a (multiline) string with information about the given dataset.
}
