% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_S4_values.R
\name{get_S4_values}
\alias{get_S4_values}
\title{Get Values From S4 Object}
\usage{
get_S4_values(S4Object)
}
\arguments{
\item{S4Object}{A S4 object or a list only containing S4 objects.}
}
\value{
A list
}
\description{
This function is used to get the slot values from S4 objects conveniently. In
addition, a list only containing one or more S4 subjects is supported.
}
