% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_split_info}
\alias{get_split_info}
\title{Summarize Dataset Split According to Trial Period}
\usage{
get_split_info(data, config)
}
\arguments{
\item{data}{\code{data.table} with the study data}

\item{config}{global config object

\code{config$first_period_end} is the last timepoint in the first trial period.
\code{config$time_variable} is the name of the variable containing timepoints in
\code{data}.
\code{config$subject_variable} is the name of the variable that identifies
subjects in \code{data}.
\code{config$group_variable} is the name of the group variable in \code{data}.}
}
\value{
printable string with dataset info
}
\description{
Create a (multiline) string with information about the given dataset.
}
