% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRootPath.R
\name{getRootPath}
\alias{getRootPath}
\title{Return the root path for the MND simulation as \code{character}.}
\usage{
getRootPath()
}
\description{
\code{getRootPath} returns the root path for the MND simulation.

It searches, successively upon failure, the path in the environment variable
of the system called "MNO_SIMULATION_PATH" and in the mno_simulation.ini file
stored in the userprofile direcotry.

If it doesn't find the path, it stops the execution and returns an error.
}
\details{
\code{character} with the name of the root path.
}
\examples{
getRootPath()



}
