% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5files.R
\name{h5_to_r}
\alias{h5_to_r}
\alias{.h5_to_r}
\title{Read An H5 File into R}
\usage{
h5_to_r(file_path)

.h5_to_r(h5file)
}
\arguments{
\item{file_path}{The file path of the candidate .h5 file}

\item{h5file}{An h5file}
}
\value{
An R object
}
\description{
Read An H5 File into R
}
\examples{
# Generate a list
x <- list("length" = 2,
          "integer" = 2L,
          "logical" = TRUE)

# Set a path
file_path <- tempfile(fileext = ".h5")
# Write an h5 file
write_h5files(x, file_path = file_path)
# Read the h5 file into R
r_object <- h5_to_r(file_path)
}
