% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{harmonize_period_times}
\alias{harmonize_period_times}
\title{Establish Equally Named Timepoints in Both Trial Periods}
\usage{
harmonize_period_times(data, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{config}{\code{list} with further arguments}
}
\value{
input data with timepoints renamed according to the configuration.
}
\description{
This is required for the implementation of multivariate prioritized gpc.
}
