% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hs_gene_length}
\alias{hs_gene_length}
\title{Gene Length of Homo sapiens}
\format{
A data frame with 25457 rows and 4 variables:
\describe{
\item{\code{gene_id}}{The entrez id of the genes}
\item{\code{length}}{Gene length of all exons}
\item{\code{symbol}}{Gene symbol}
\item{\code{ensembl_id}}{Ensembl id of genes}
}
}
\usage{
hs_gene_length
}
\description{
A dataframe containing the gene length of Homo sapiens.
}
\keyword{datasets}
