% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{is_define_parameters}
\alias{is_define_parameters}
\title{Check Class of Parameter Objects}
\usage{
is_define_parameters(parameters_list)
}
\arguments{
\item{parameters_list}{A list consists of defined parameter.}
}
\description{
Make sure all the parameters are generated by \code{param_numeric},
\code{param_character}, \code{param_Boolean} or \code{param_integer} function.
}
\seealso{
\code{\link[=param_numeric]{param_numeric()}}, \code{\link[=param_character]{param_character()}}, \code{\link[=param_Boolean]{param_Boolean()}}, \code{\link[=param_integer]{param_integer()}}
}
