% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs_debias.r
\name{lhs_debias}
\alias{lhs_debias}
\title{Perform Debiasing Regression for Left-Hand-Side (LHS) Simulation}
\usage{
lhs_debias(train, test, suppressWarnings = FALSE)
}
\arguments{
\item{train}{A data frame representing the training set, with columns Yllm, Yhuman, and V.}

\item{test}{A data frame representing the test set, with columns Yllm and V.}

\item{suppressWarnings}{Logical. If TRUE, suppresses warnings during model fitting.}
}
\value{
A list containing:
  \item{coef_Ytilde_V}{Debiased coefficients from the regression of Ytilde on V.}
  \item{other}{Intermediate regression, namely error ~ V.}
}
\description{
This function estimates the error between two variables (Yllm and Yhuman) and performs a regression to debias the outcome (Ytilde).
}
