% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs_simulate.r
\name{lhs_simulate}
\alias{lhs_simulate}
\title{Simulate Left-Hand-Side (LHS) Regressions Based on a Specified Combination}
\usage{
lhs_simulate(
  combination,
  data,
  N,
  B,
  n_samples,
  type_boot = c("bayesian", "nonparametric"),
  rds_dir = NULL
)
}
\arguments{
\item{combination}{A list specifying the model and variables to use in the simulation.}

\item{data}{A data frame containing the full dataset.}

\item{N}{Integer. The number of simulations to perform.}

\item{B}{Integer. The number of bootstrap samples to generate.}

\item{n_samples}{Integer. The number of samples to draw for each simulation.}

\item{type_boot}{Character. Specifies the type of bootstrap ("bayesian" or "nonparametric").}

\item{rds_dir}{Optional character. If provided, the results will be saved as RDS files in the specified directory.}
}
\value{
A data frame summarizing the results of all simulations or a path to the saved RDS file.
}
\description{
This function performs simulations for LHS regressions based on combinations of variables and models, performing bootstrap sampling to estimate uncertainty.
}
