% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_docker_container.R
\name{list_docker_container}
\alias{list_docker_container}
\title{List the Containers Existed in Docker}
\usage{
list_docker_container(...)
}
\arguments{
\item{...}{Nothing to input}
}
\value{
A tibble
}
\description{
Utils function of getting the containers information in R.
}
\examples{
containers <- list_docker_container()
}
