% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_trees.R
\name{make_trees}
\alias{make_trees}
\title{Convert An Expression Matrix into Tree Format}
\usage{
make_trees(ref_data, group, is_Newick = TRUE, is_parenthetic = FALSE)
}
\arguments{
\item{ref_data}{A Matrix. The gene expression matrix where every row represents
a cell and every column represents a gene.}

\item{group}{A Vector. The group information of the cells in ref_data}

\item{is_Newick}{If TRUE, return the Newick format.}

\item{is_parenthetic}{If TRUE, return the parenthetic format.}
}
\value{
A file of parenthetic format or Newick format
}
\description{
This function is used to convert the gene expression matrix into the tree
format.
}
\examples{
ref_data <- matrix(rpois(n = 2500, lambda = 2), nrow = 50)
rownames(ref_data) <- paste0("cell_", 1:ncol(ref_data))
colnames(ref_data) <- paste0("gene_", 1:nrow(ref_data))
group <- c(rep("A", 15), rep("B", 5), rep("C", 20), rep("D", 10))
tree_format <- make_trees(ref_data, group = group)
}
