% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_definition.R
\name{manuscript_definition}
\alias{manuscript_definition}
\title{Manuscript Information}
\usage{
manuscript_definition(
  title,
  doi,
  journal = NULL,
  date = NULL,
  peer_review = TRUE
)
}
\arguments{
\item{title}{Manuscript title.}

\item{doi}{Doi number.}

\item{journal}{Journal name.}

\item{date}{Publication date.}

\item{peer_review}{Has been peer reviewed?}
}
\value{
A list.
}
\description{
Manuscript Information
}
