% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm_gene_length}
\alias{mm_gene_length}
\title{Gene Length of Mus musculus}
\format{
A data frame with 24115 rows and 4 variables:
\describe{
\item{\code{gene_id}}{The entrez id of the genes}
\item{\code{length}}{Gene length of all exons}
\item{\code{symbol}}{Gene symbol}
\item{\code{ensembl_id}}{Ensembl id of genes}
}
}
\usage{
mm_gene_length
}
\description{
A dataframe containing the gene length of Mus musculus.
}
\keyword{datasets}
