% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_model.R
\name{model_predict}
\alias{model_predict}
\title{Establish A Model And Predict Cell Identity Using DEGs}
\usage{
model_predict(data, group, de_genes, method, verbose)
}
\arguments{
\item{data}{A matrix with cells on columns and genes on rows.}

\item{group}{Group(or cluster) assignment of every cells in columns of matrix.}

\item{de_genes}{A character vector of DEGs.}

\item{method}{The method to establish the model. SVM, Decision tree or RF (Random Forest).}

\item{verbose}{Whether the process massages are returned.}
}
\description{
Establish A Model And Predict Cell Identity Using DEGs
}
