% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{na_count}
\alias{na_count}
\title{Count Number of Failed Hypothesis Tests}
\usage{
na_count(results_df)
}
\arguments{
\item{results_df}{\code{data.frame} with columns of individual test results (p-values)}
}
\value{
list of sums with names equal to the input's column names
}
\description{
based on data.frame of test results that contains \code{NA} for each failed test
}
