% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_similarity.R
\name{node_height_dis}
\alias{node_height_dis}
\title{Calculate the Discrepency of Nodes' Height}
\usage{
node_height_dis(ref_tree, sim_tree)
}
\arguments{
\item{ref_tree}{Phylo tree derived from reference data}

\item{sim_tree}{Phylo tree derived from simulated data}
}
\value{
An RMSE value
}
\description{
Calculate the Discrepency of Nodes' Height
}
