% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization_simutils}
\alias{normalization_simutils}
\title{FPKM or TPM Normalization Method in simutils}
\usage{
normalization_simutils(
  data,
  norm_method = "FPKM",
  gene_length = NULL,
  species = "human"
)
}
\arguments{
\item{data}{A matrix with cells on columns and genes on rows.}

\item{norm_method}{FPKM or TPM}

\item{gene_length}{A numeric vector of gene lengths. If NULL, the gene length data will be imported by \link{hs_gene_length} or \link{mm_gene_length} function and users must set \code{species} parameters below.}

\item{species}{\code{human} or \code{mouse}.}
}
\description{
FPKM or TPM Normalization Method in simutils
}
