% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_Boolean}
\alias{param_Boolean}
\title{Define Logical Parameters}
\usage{
param_Boolean(id, type = "logical", default, description)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be a logical value.}

\item{default}{The default value. TRUE or FALSE.}

\item{description}{The description information of this parameter.}
}
\value{
A list.
}
\description{
Define Logical Parameters
}
\examples{
verbose <- param_Boolean(id = "verbose",
                         default = TRUE,
                         description = "Whether to return the information during the process")
}
