% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_character}
\alias{param_character}
\title{Define Character Parameters}
\usage{
param_character(id, type = "character", default, alternatives, description)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be characters.}

\item{default}{The default values. Length to one or more.}

\item{alternatives}{Options that can be chosen.}

\item{description}{The description information of this parameter.}
}
\value{
A list.
}
\description{
Define Character Parameters
}
\examples{
dimension_method <- param_character(id = 'dimension_method',
                                    default = "UMAP",
                                    alternatives = c("UMAP", "TSNE", "PCA"),
                                    description = " A character string specifying the algorithm to use for dimensionality reduction.")
}
