% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_inference}
\alias{param_inference}
\title{Define Inference Dataset Parameter}
\usage{
param_inference(
  id = NULL,
  type = NULL,
  alternatives = c("matrix", "SingleCellExperiment"),
  description = NULL,
  process = "simulation"
)
}
\arguments{
\item{id}{A parameter name. Default is NULL, which means no reference matrix input.}

\item{type}{The type of parameter. Default is NULL. See \code{alternatives}.}

\item{alternatives}{Alternative options.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}
}
\value{
A list.
}
\description{
Define Inference Dataset Parameter
}
\examples{
matrix_input_method <- param_inference(
id = 'matrix_input',
type = "SingleCellExperiment",
process = "estimation",
description = "The matrix to use for parameters estimating."
)
}
