% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_integer}
\alias{param_integer}
\title{Define Integer Parameters}
\usage{
param_integer(
  id,
  type = "integer",
  default = NULL,
  lower = NULL,
  upper = NULL,
  border = TRUE,
  force = FALSE,
  description = NULL,
  process = "simulation",
  function_name
)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be an integer.}

\item{default}{The default value.}

\item{lower}{The lowest range of this parameter.}

\item{upper}{The highest range of this parameter.}

\item{border}{Boolean. Whether the default value can get to the range border.}

\item{force}{Logical, whether this parameter must be needed or not.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}

\item{function_name}{The function name where the parameter exists.}
}
\value{
A list.
}
\description{
Define an integer parameter.
}
\examples{
group_num <- param_integer(
id = "group_num",
default = 1L,
lower = 1L,
upper = 50L,
border = TRUE,
process = "estimation",
description = "How many groups to be simulated"
)
}
