% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_reference}
\alias{param_reference}
\title{Define Reference Dataset Parameters}
\usage{
param_reference(
  id = NULL,
  type = NULL,
  default = NULL,
  alternatives = c("matrix", "SingleCellExperiment", "dataframe"),
  force = FALSE,
  description = NULL,
  process = "simulation",
  function_name
)
}
\arguments{
\item{id}{A parameter name. Default is NULL, which means no reference matrix input.}

\item{type}{The type of parameter. Default is NULL. See \code{alternatives}.}

\item{default}{In simulation, it means the reference dataset where methods learn the parameters from. Default is NULL.}

\item{alternatives}{Alternative options.}

\item{force}{Logical, whether this parameter must be needed or not.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}

\item{function_name}{The function name where the parameter exists.}
}
\value{
A list.
}
\description{
Define Reference Dataset Parameters
}
\examples{
matrix_input_method <- param_reference(
id = 'matrix_input',
type = "SingleCellExperiment",
process = "estimation",
description = "The matrix to use for parameters estimating."
)
}
