% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_vector}
\alias{param_vector}
\title{Define Vector Parameters}
\usage{
param_vector(
  id,
  type = "vector",
  default = NULL,
  description = NULL,
  process = "simulation"
)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter.}

\item{default}{The default values. Length to one or more.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}
}
\value{
A list.
}
\description{
Define Vector Parameters
}
\examples{
path_method <- param_vector(
id = 'path',
default = c(0,1,2,3),
process = "simulation"
)
}
