% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameter_sets}
\alias{parameter_sets}
\title{Generate Parameter sets}
\usage{
parameter_sets(...)
}
\arguments{
\item{...}{Defined lists of parameters.}
}
\value{
A list contains the parameters of estimating process and simulation process
}
\description{
Collect and return all defined parameters.
}
\examples{
param_set <- parameter_sets(
param_character(id = 'dimension_method',
                default = "UMAP",
                alternatives = c("UMAP", "TSNE", "PCA"),
                force = TRUE,
                description = "The algorithm to use for dimensionality reduction."),
param_Boolean(id = "verbose",
              default = TRUE,
              process = 'estimation',
              description = "Whether to return the information during the process"),
param_integer(id = "group_num",
              default = 1L,
              lower = 1L,
              upper = 50L,
              border = TRUE,
              description = "How many groups to be simulated"),
param_numeric(id = "probability",
              default = 0.5,
              lower = 0,
              upper = 1,
              border = TRUE,
              force = TRUE,
              process = 'estimation',
              description = "The probability to select the right number")
)
}
\seealso{
\code{\link[=param_numeric]{param_numeric()}}, \code{\link[=param_character]{param_character()}}, \code{\link[=param_Boolean]{param_Boolean()}}, \code{\link[=param_integer]{param_integer()}}
}
