% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_DE_analysis.R
\name{perform_DEA}
\alias{perform_DEA}
\title{Perform Differential Expression Analysis
This function is used to perform DEA for single-cell RNA-seq data by QLF model in edgeR package including the cellular detection rate.}
\usage{
perform_DEA(data, group)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{group}{Group(or cluster) assignment of every cells in columns of matrix.}
}
\value{
A dataframe of DEA result.
}
\description{
Perform Differential Expression Analysis
This function is used to perform DEA for single-cell RNA-seq data by QLF model in edgeR package including the cellular detection rate.
}
\references{
Soneson C, Robinson M D. Bias, robustness and scalability in single-cell differential expression analysis. Nature methods, 2018, 15(4): 255-261. \url{https://doi.org/10.1038/nmeth.4612}
}
