% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{print_config_to_stderr}
\alias{print_config_to_stderr}
\title{Pretty-Print Program Configuration.}
\usage{
print_config_to_stderr(options, config)
}
\arguments{
\item{options}{List with command line options. All except from 'help' will
be printed}

\item{config}{List with config parameters. Only selected elements will be
printed}
}
\description{
Pretty-Print Program Configuration.
}
