% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{print_data_info_to_stderr}
\alias{print_data_info_to_stderr}
\title{Summarize Dataset Split According to Trial Period}
\usage{
print_data_info_to_stderr(data, config)
}
\arguments{
\item{data}{\code{data.table} with the study data}

\item{config}{global config object

\code{config$first_period_end} is the last timepoint in the first trial period.
\code{config$time_variable} is the name of the variable containing timepoints in
\code{data}.
\code{config$subject_variable} is the name of the variable that identifies
subjects in \code{data}.
\code{config$group_variable} is the name of the group variable in \code{data}.}
}
\description{
Print each trial period info as a separate data.frame.
}
