% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{print_list_to_stderr}
\alias{print_list_to_stderr}
\title{Pretty-Print a Flat List to Stderr}
\usage{
print_list_to_stderr(l, header)
}
\arguments{
\item{l}{The list to print}

\item{header}{Title of the list. Will be printed above}
}
\description{
Pretty-Print a Flat List to Stderr
}
