% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv}
\alias{read_csv}
\title{Read a csv file according to a schema specified in an xml file.}
\usage{
read_csv(xmlFileName, csvFileName)
}
\arguments{
\item{xmlFileName}{File name of the xml file containing the schema for the
csv file.}

\item{csvFileName}{File name of the csv file to validate.}
}
\description{
\code{read_csv} reads a csv file assigning column names and
column types according to the specifications in an xml file.
}
\details{
Return a \code{data.table} with the contents of the csv file.
}
\examples{
rootPath <- file.path(system.file(package = "simutils"), "extdata")

# simulation file
csv_fn <- file.path(rootPath, "output_files", "antennas.csv")
xml_fn <- file.path(rootPath, "metadata/output_files", "antennas_dict.xml")
dt <- read_csv(xml_fn, csv_fn)
str(dt)

# coverage file
csv_fn <- file.path(rootPath, "output_files", "AntennaCells_MNO1.csv")
xml_fn <- file.path(rootPath, "metadata/output_files", "antennaCells_dict.xml")
dt <- read_csv(xml_fn, csv_fn)
str(dt)

}
