% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_data}
\alias{read_data}
\title{Read in Diacerin Study Dataset}
\usage{
read_data(filename, config)
}
\arguments{
\item{filename}{path to the tab-separated dataset file}

\item{config}{\code{list} with further arguments}
}
\value{
the preprocessed dataset as \code{data.table}
}
\description{
The dataset is read in as \code{data.table}. This is important as many of the
other functions require this specific data type. The dataset is supposed to
consist of tab-separated values (tsv). Its rows will be sorted according
to \code{config$subject_variable} and \code{config$time_variable}. For this purpose,
the time variable will be casted to numeric.
}
