% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simutils.R
\name{read_sim_dir}
\alias{read_sim_dir}
\title{Read simulation directory}
\usage{
read_sim_dir(sim_dir)
}
\arguments{
\item{sim_dir}{A character string representing the directory path of the simulation. \code{dirname(sim_dir)} must exist, but \code{basename(sim_dir)} will be created if necessary.}
}
\value{
Returns list of simulation results.
}
\description{
Reads and returns results existing in a simulation directory.
}
