% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWKT_as_sfc.R
\name{readWKT_as_sfc}
\alias{readWKT_as_sfc}
\title{Read string in WKT format and convert it into an sf geometry list column.}
\usage{
readWKT_as_sfc(
  x,
  ...,
  crs = NA_integer_,
  EWKB,
  spatialite,
  pureR,
  precision = 0,
  forceMulti = FALSE,
  GeoJSON
)
}
\arguments{
\item{x}{Character string of WKT}

\item{...}{further arguments}

\item{crs}{integer or character; coordinate reference system for the geometry
as in function \code{sf::st_as_sfc}}

\item{EWKB}{logical; if TRUE, parse as EWKB
(extended WKB; PostGIS: ST_AsEWKB), otherwise as ISO WKB
(PostGIS: ST_AsBinary)}

\item{spatialite}{logical; if TRUE, WKB is assumed to be in the spatialite
dialect, see https://www.gaia-gis.it/gaia-sins/BLOB-Geometry.html; this is
only supported in native endian-ness (i.e., files written on system with the
same endian-ness as that on which it is being read).}

\item{pureR}{logical; if TRUE, use only R code, if FALSE, use compiled (C++)
code; use TRUE when the endian-ness of the binary differs from the host
machine (.Platform$endian).}

\item{precision}{precision value; see st_as_binary}

\item{forceMulti}{logical; if TRUE, force coercion into MULTIPOLYGON or
MULTILINE objects, else autodetect}

\item{GeoJSON}{logical; if TRUE, try to read geometries from GeoJSON text
strings geometry, see st_crs()}
}
\description{
Read the input WKT string, parse it, and convert it into an sf geometry list
column.
}
\examples{
txt <- readLines(
   system.file("extdata/input_files", "map.wkt", package = "simutils"))
map_polygon_sfc <- readWKT_as_sfc(txt)
class(map_polygon_sfc)

}
\seealso{
\code{\link[sf]{st_as_sfc}}
}
