% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_topics.r
\name{recode_topics}
\alias{recode_topics}
\title{Recode Topics to Predefined List with "Other" for Non-Matching Topics}
\usage{
recode_topics(x, topics)
}
\arguments{
\item{x}{A vector of topics to be recoded.}

\item{topics}{A vector of predefined topics.}
}
\value{
A recoded factor vector with non-matching topics labeled as "Other".
}
\description{
This function recodes a vector of topics using a provided list of topics and converts non-included topics to "Other".
}
