% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rejection_rate}
\alias{rejection_rate}
\title{Compute Null Hypothesis Rejection Rate}
\usage{
rejection_rate(results_df, alpha)
}
\arguments{
\item{results_df}{\code{data.frame} with columns of individual test results (p-values)}

\item{alpha}{type-I error rate}
}
\value{
list of rejection rates with names equal to the input's column names
}
\description{
based on data.frame of test results
}
