% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhs_debias.r
\name{rhs_debias}
\alias{rhs_debias}
\title{Perform Debiasing Regression for Right-Hand-Side (RHS) Simulation}
\usage{
rhs_debias(train, test, suppressWarnings = FALSE)
}
\arguments{
\item{train}{A data frame representing the training set, with columns V, Yhuman, Yllm, and possibly weights.}

\item{test}{A data frame representing the test set, with columns V, Yhuman, Yllm, and possibly weights.}

\item{suppressWarnings}{Logical. If TRUE, suppresses warnings during model fitting.}
}
\value{
A list containing:
  \item{coef_nu_Yllm}{Coefficients from the regression of nu on Yllm.}
  \item{coef_Vtilde_Ytilde}{Debiased coefficients from the regression of Vtilde on Ytilde.}
  \item{other}{A list of intermediate regressions, including V ~ Yhuman, V ~ Yllm, and Yhuman ~ Yllm.}
}
\description{
This function performs a debiasing regression for RHS variables, regressing the variable of interest (Vtilde) on the predicted values (Ytilde).
The function calculates intermediate regressions and predicts debiased coefficients for the test data.
}
