% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DockerImage.R
\name{run_DockerImage}
\alias{run_DockerImage}
\title{Run the simulation software using the docker image.}
\usage{
run_DockerImage(
  input_folder,
  simulationCFGFile,
  mapFile,
  personsCFGFile,
  antennasCFGFile,
  output_folder
)
}
\arguments{
\item{input_folder}{Absolute path of the folder which contains the input data.}

\item{simulationCFGFile}{Name of the simulation configuration file.}

\item{mapFile}{Name of the map file.}

\item{personsCFGFile}{Name of the persons' configuration file.}

\item{antennasCFGFile}{Name of the antennas' configuration file.}

\item{output_folder}{Name of the output folder indicated in the
simulation's configuration file.}
}
\description{
\code{run_DockerImage} runs the simulation software using
the docker image.
}
\details{
Return invisible NULL after executing the docker image.
}
\examples{
rootPath <- system.file(package = 'simutils')
input_folder      <- file.path(rootPath, "extdata/input_files")
simulationCFGFile <- 'simulation.xml'
mapFile           <- 'map.wkt'
personsCFGFile    <- 'persons.xml'
antennasCFGFile   <- 'antennas.xml'
output_folder     <- file.path(Sys.getenv('HOME'), 'example_docker')
\dontrun{
# run only after checking your installation of docker
run_DockerImage(
  input_folder = input_folder,
  simulationCFGFile = simulationCFGFile,
  mapFile = mapFile,
  personsCFGFile = personsCFGFile,
  antennasCFGFile = antennasCFGFile,
  output_folder = output_folder)
}
}
