% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExeFile.R
\name{runExeFile}
\alias{runExeFile}
\title{Run an MND simulation using the simulator.exe file.}
\usage{
runExeFile(
  path_to_exe,
  input_folder,
  simulationCFGFile,
  mapFile,
  personsCFGFile,
  antennasCFGFile,
  outputFolder
)
}
\arguments{
\item{path_to_exe}{character vector with the absolute path to the executable
file of the simulator of mobile network event data.}

\item{input_folder}{absolute path where the simulation input files are.}

\item{simulationCFGFile}{name of the configuration simulation file
(input file).}

\item{mapFile}{name of the map file (input file).}

\item{personsCFGFile}{name of the persons file (input file).}

\item{antennasCFGFile}{name of the antennas file (input file).}

\item{outputFolder}{Absolute path where the output files are going to be saved.}
}
\description{
\code{runExeFile} reads the simulator path to the executable
file of the simulator of mobile network event data and the input file
parameters to execute the simulator.exe from this path.

If the file doesn't exist, it returns an error.
}
\details{
Return invisible \code{NULL} after placing output files of the
simulation in the output folder.
}
\examples{
rootPath <- system.file(package = "simutils")
runExeFile(
 path_to_exe       = file.path(rootPath, 'bin'),
 input_folder      = file.path(rootPath, "extdata/input_files"),
 simulationCFGFile = "simulation.xml",
 mapFile           = "map.wkt",
 personsCFGFile    = "persons.xml",
 antennasCFGFile   = "antennas.xml",
 outputFolder      = Sys.getenv('R_USER'))
  
  
}
