% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sanity_check}
\alias{sanity_check}
\title{Check User Input}
\usage{
sanity_check(options, config)
}
\arguments{
\item{options}{\code{list} with user-provided command line arguments}

\item{config}{\code{list} that contains valid options to compare with}
}
\description{
Perform simple sanity checks. Throw error if invalid user input is detected.
}
