% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{scgan_data_conversion}
\alias{scgan_data_conversion}
\title{Convert R Matrix to Python h5ad File}
\usage{
scgan_data_conversion(data, data_id, save_to_path, verbose)
}
\arguments{
\item{data}{A gene expression matrix with cell on columns and gene on rows.}

\item{data_id}{Tha data name to be output.}

\item{save_to_path}{The save path on local device.}

\item{verbose}{If messages are returned during the process.}
}
\description{
Convert R Matrix to Python h5ad File
}
