% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se.boot}
\alias{se.boot}
\title{Standard Errors from Bootstrap Results}
\usage{
\method{se}{boot}(x)
}
\arguments{
\item{x}{A bootstrap object created by the `boot` function.}
}
\value{
A vector of standard errors based on the bootstrap results.
}
\description{
This method calculates the standard errors from bootstrapped samples.
}
