% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se.lm}
\alias{se.lm}
\title{Standard Errors for a Linear Model}
\usage{
\method{se}{lm}(x)
}
\arguments{
\item{x}{A linear model object (of class `lm`).}
}
\value{
A vector of standard errors for the coefficients in the model.
}
\description{
This method extracts standard errors from a fitted linear model (`lm` object).
}
