% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se.mlm}
\alias{se.mlm}
\title{Standard Errors for a Multivariate Linear Model}
\usage{
\method{se}{mlm}(x)
}
\arguments{
\item{x}{A multivariate linear model object (of class `mlm`).}
}
\value{
A named vector of standard errors for each coefficient, separated by response variable.
}
\description{
This method extracts standard errors for each response in a multivariate linear model (of class `mlm`).
}
