% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se}
\alias{se}
\title{Generic Function for Standard Errors}
\usage{
se(x)
}
\arguments{
\item{x}{An object for which standard errors are to be extracted.}
}
\value{
A vector or matrix of standard errors, depending on the object type.
}
\description{
This is a generic function to extract standard errors from a variety of model objects. It dispatches methods based on the class of the input.
}
