% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se.robust}
\alias{se.robust}
\title{Standard Errors for Robust Objects}
\usage{
\method{se}{robust}(x)
}
\arguments{
\item{x}{An object of class `robust`.}
}
\value{
A vector of robust standard errors.
}
\description{
This method extracts standard errors from a robust object created by the `robust` function.
}
