% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.r
\name{se.summary.lm}
\alias{se.summary.lm}
\title{Standard Errors for a summary object}
\usage{
\method{se}{summary.lm}(x)
}
\arguments{
\item{x}{A summary object (of class `summary.lm`).}
}
\value{
A vector of standard errors for the coefficients in the model.
}
\description{
This method extracts standard errors from summary object of a linear model (`summary.lm` object).
}
