% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_manipulation.R
\name{set_param}
\alias{set_param}
\title{Set Parameter Values}
\usage{
set_param(method, param_list, reset)
}
\arguments{
\item{method}{The methods where you want to change the parameters from. One or more methods are supported.}

\item{param_list}{A list generated by \code{\link[simmethods:get_method]{simmethods::get_method()}}.}

\item{reset}{e.g. c("nCells:2000", "nGenes:20000", "nBatches:10") or a list for setting parameters in more methods. See examples.}
}
\value{
An update list of parameters
}
\description{
Set Parameter Values
}
\seealso{
\code{\link[simmethods:get_method]{simmethods::get_method()}}
}
