% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_manipulation.R
\name{set_parameters}
\alias{set_parameters}
\title{Set Paramters}
\usage{
set_parameters(parameters, other_prior, method)
}
\arguments{
\item{parameters}{The estimation result generated by \code{simpipe::estimate_parameters()}}

\item{other_prior}{A list with names of certain parameters. Some methods need
extra parameters to execute the estimation step, so you must input them. In
simulation step, the number of cells, genes, groups, batches, the percent of
DEGs and other variables are usually customed, so before simulating a dataset
you must point it out.}

\item{method}{Method name.}
}
\description{
Set new parameter values via prior infomation
}
