% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_signal_thresh.R
\name{set_signal_thresh}
\alias{set_signal_thresh}
\title{Create RSS_thresh and SDM_thresh from the SimData object.}
\usage{
set_signal_thresh(simData)
}
\arguments{
\item{simData}{list of different elements from the simulation.}
}
\description{
This function computes the RSS_thresh and the SDM_thresh of each antenna
according to the minimum values (Smin and Qmin), contained in
the simData$network object, related with RSS and SDM attributes from
the simData$grid object. It returns the RSS_thresh and SDM_thresh as
new attributes of simData$grid.
}
\examples{
filename_map      <- c(
 xml= system.file("extdata/input_files", "map.xml", package = "simutils"),
 xsd= '')
  
filename_network  <- c(
 csv= system.file("extdata/output_files/antennas.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/antennas_dict.xml", package = "simutils"))
                   
filename_signal <- c(
 csv= system.file("extdata/output_files/SignalMeasure_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/SignalMeasure_dict.xml", package = "simutils"))
                 
filename_coverage <- c(
 csv= system.file("extdata/output_files", "AntennaCells_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/AntennaCells_dict.xml", package = "simutils"))

filename_events <- c(
 csv= system.file("extdata/output_files/AntennaInfo_MNO_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/events_dict.xml", package = "simutils"))
                       
filename_grid <- c(
  csv= system.file("extdata/output_files/grid.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/grid_dict.xml", package = "simutils")) 

filename_individ <- c(
  csv= system.file("extdata/output_files/persons.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/persons_dict.xml", package = "simutils"))   
                       
filenames <- list(
  map                = filename_map,
  network_parameters = filename_network,
  signal             = filename_signal,
  events             = filename_events,
  coverage_cells     = filename_coverage,
  grid               = filename_grid,
  individuals        = filename_individ)
  
simData <- simutils::read_simData(filenames, crs = 2062)
simData <- simutils::compute_sdm(simData)
set_signal_thresh(simData)

}
