% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simutils.R
\name{sim_across}
\alias{sim_across}
\title{Simulate across settings}
\usage{
sim_across(
  sim_fn,
  sim_dir = file.path(getwd(), time2id()),
  n_reps = 1,
  seed = 1,
  n_cores = 1,
  mc_method = c("basic", "redundant"),
  debug = 1,
  ...
)
}
\arguments{
\item{sim_fn}{A simulation function.}

\item{sim_dir}{A character string representing the directory path of the simulation. \code{dirname(sim_dir)} must exist, but \code{basename(sim_dir)} will be created if necessary.}

\item{n_reps}{A positive integer representing the number of repetitions of \code{sim_fn} for each combination of settings.}

\item{seed}{A list containing the simulation settings to be saved or validated.}

\item{n_cores}{A positive integer representing the number of cores to use for execution. Set to -1 to automatically detect the number of cores.}

\item{mc_method}{A character value representing the multicore execution method. \code{"basic"} allocates one core to each unique combination of settings, whereas \code{"redundant"} will give every core an opportunity to work on each setting combination.}

\item{debug}{An integer for debugging level. Defaults to 1.}

\item{...}{Additional settings passed to \code{sim_fn}. \code{sim_fn} will be executed over each unique combination of settings. Optionally, a \code{sim_grid} data.frame object can be provided with the desired setting combinations.}
}
\value{
Returns list of simulation results.
}
\description{
Simulates a simulation function across combinations of settings.
}
\examples{
## Example simulation function
sim_fn <- function(n, mean, sd){
  Sys.sleep(0.5)
  x <- rnorm(n = n, mean = mean, sd = sd)
  data.frame(mean = mean(x),
             sd = sd(x))
}

## Example execution using ... and single core
res1 <- sim_across(sim_fn = sim_fn, sim_dir = "inst/example/example1",
                   n_reps = 16, seed = 1, n_cores = 1,
                   n = seq_len(2) * 10, mean = seq_len(2), sd = seq_len(2),
                   debug = 1)

## Example execution using sim_grid and basic multicore
sim_grid <- expand.grid(n = seq_len(2) * 10,
                        mean = seq_len(2),
                        sd = seq_len(2))

res2 <- sim_across(sim_fn = sim_fn, sim_dir = "inst/example/example2",
                   n_reps = 16, seed = 1, n_cores = -1, mc_method = "basic",
                   sim_grid = sim_grid,
                   debug = 1)

all.equal(res1, res2, check.attributes = FALSE)

## Example execution using ... and redundant multicore
res3 <- sim_across(sim_fn = sim_fn, sim_dir = "inst/example/example3",
                   n_reps = 16, seed = 1, n_cores = -1, mc_method = "redundant",
                   n = seq_len(2) * 10, mean = seq_len(2), sd = seq_len(2),
                   debug = 1)

all.equal(res2, res3, check.attributes = FALSE)
}
