% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{start_cell}
\alias{start_cell}
\title{Define A Start Cell of A Trajectory}
\usage{
start_cell(meta_data)
}
\arguments{
\item{meta_data}{The position information of cells or Spots}
}
\value{
A cell id
}
\description{
Define A Start Cell of A Trajectory
}
