% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{subtract_baseline}
\alias{subtract_baseline}
\title{Subtract Baseline from Observations}
\usage{
subtract_baseline(data, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\description{
Target variable values are supposed to be divided into blocks of equal size.
Each block combines measurements for one subject at different points of
time. The baseline measurement is the one with the earliest time point.
}
\details{
\code{options$target} contains the name of the target variable.
\code{options$subtract} determines if the baseline observation should be
subtracted from other target observations or not (\code{TRUE}/\code{FALSE}).
\code{config$blocklength} is the number of measurements in a block that belong to
one subject.

Note that \code{data} is modified in place.
}
