% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_tests}
\alias{summarize_tests}
\title{Summarize Hypothesis Tests}
\usage{
summarize_tests(results_df, alpha)
}
\arguments{
\item{results_df}{\code{data.frame} with columns of individual test results (p-values)}

\item{alpha}{type-I error rate}
}
\value{
summary list
}
\description{
compute H0 rejection rate and count number of failed tests based on a
data.frame of test results that contains p-values (or \code{NA} for a
failed test).
}
