% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.list}
\alias{summary.list}
\title{Summary for a List of Objects}
\usage{
\method{summary}{list}(..., alpha = 0.05, z_score = TRUE)
}
\arguments{
\item{...}{A list of objects to summarize.}

\item{alpha}{The significance level for confidence intervals (default is 0.05).}

\item{z_score}{Logical indicating whether to use z-scores (default is TRUE).}
}
\value{
A data frame summarizing the objects.
}
\description{
This method generates a combined summary for a list of objects (e.g., robust or bootstrap objects).
}
