% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_cells.R
\name{synthesize_cells}
\alias{synthesize_cells}
\title{Synthesize Fake Cells}
\usage{
synthesize_cells(dataset, group = NULL, seed)
}
\arguments{
\item{dataset}{A matrix or the result generated by \code{\link[dynwrap]{wrap_expression}}}

\item{group}{A vector. Default is NULL.}

\item{seed}{Integer. A random seed.}
}
\value{
A list generated by \code{\link[dynwrap]{wrap_expression}}
}
\description{
This function is used to synthesize fake cells via linear combination when the
number of cells is not the power of 2. First, we need cell group information
where real cells are sampled from. If no group information input, we perform
k-means algorithm on the data and use \code{\link[NbClust]{NbClust}} funciton
to determin the best cluster number. Finally, we merge the synthesized and
real data as the output result.
}
\examples{
set.seed(1)
a <- matrix(rpois(n = 2500, lambda = 2), nrow = 50)
rownames(a) <- paste0("cell_", 1:ncol(a))
colnames(a) <- paste0("gene_", 1:nrow(a))
dataset_ref <- dynwrap::wrap_expression(
  counts = a,
  expression = log2(a+1)
)
result <- syn_cell(dataset = a, seed = 2)
}
