% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{test_h0}
\alias{test_h0}
\title{Perform Hypothesis Test}
\usage{
test_h0(data, period, target, alpha)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{period}{study period (either 1 or 2)}

\item{target}{name of the target variable}

\item{alpha}{type-I error rate}
}
\value{
the test result (TRUE if H0 is rejected, FALSE otherwise)
}
\description{
Split the dataset accorind to the soecified period and perform a hypothesis
test.
}
