% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGs_manipulation.R
\name{test_uni_distribution}
\alias{test_uni_distribution}
\title{Test if the p-value is Uniformly Distributed}
\usage{
test_uni_distribution(PValue)
}
\arguments{
\item{PValue}{The p-value vector.}
}
\description{
Test if the p-value is Uniformly Distributed
}
\examples{
pvalue <- runif(1000)
result <- test_uni_distribution(pvalue)
}
