% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmacthed_GPC_github.R
\name{test_unmatched_gpc}
\alias{test_unmatched_gpc}
\title{Perform Hypothesis Test}
\usage{
test_unmatched_gpc(data, target, type, alpha, side)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{target}{name of the target variable}

\item{type}{the target variable (binary or count)}

\item{alpha}{type-I error rate}

\item{side}{1 or 2 for one- or two-sided test}
}
\value{
the test result (TRUE if H0 is rejected, FALSE otherwise)
}
\description{
unmatched GPC test assuming larger values are preferred
}
