% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEGs_manipulation.R
\name{true_DEGs_proportion}
\alias{true_DEGs_proportion}
\title{Proportion of Matched DEGs}
\usage{
true_DEGs_proportion(
  sim_data,
  group,
  group_combn,
  sim_DEGs,
  DEA_method,
  verbose = TRUE
)
}
\arguments{
\item{sim_data}{The simulated dataset to be evaluated.}

\item{group}{Group(or cluster) assignment of every cells in columns of matrix.}

\item{group_combn}{The combinations of two groups generated by \code{\link[utils:combn]{utils::combn()}} function.}

\item{sim_DEGs}{A list with \code{xxxvsxxx} format of names containing the DEGs of different paired groups.}

\item{DEA_method}{The method used for differential expression analysis. edgeRQLF, edgeRQLFDetRate, MASTcpmDetRate,
MASTcpm, limmatrend and limmavoom.}

\item{verbose}{Whether the process massages are returned.}
}
\description{
Proportion of Matched DEGs
}
