% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_SimInput.R
\name{update_SimInput}
\alias{update_SimInput}
\title{Update simulation xml input files.}
\usage{
update_SimInput(xmlSimInput, newParam, xsdName, newFileName = NULL)
}
\arguments{
\item{xmlSimInput}{Object of class \code{xml_document} from package
\code{xml2} or of class \code{character} specifying the path of the
simulation input xml file to update.}

\item{newParam}{\code{list} with the new values of the parameters to be
updated}

\item{xsdName}{\code{character} of length 1 with the name of xsd file
containing the schema definition of the xml input file.}

\item{newFileName}{\code{character} of length 1 with the name of the updated
xml input file.}
}
\description{
This function takes an xml input file for the simulator and
updates it according to the set of new parameters specified in the input
parameter list \code{newParam}. The schema definition of the xml file is also
required as an input parameter to validate the structure of the output file.
}
\details{
Return an \code{xml-document} object of package \code{xml2} with
the parameters updated according to the specified input parameters. If
\code{newFileName} is specified, this object is written in the corresponding
xml file.
}
\examples{
 
rootPath <- system.file(package = 'simutils')

xmlSimInput  <- file.path(rootPath, 'extdata/input_files', 'simulation.xml')
newParam.lst <- list(
 end_time = 11,
 movement_pattern = structure(list(
    manhattan_grid = structure(
        list(x_step = 40, y_step = 40, x_origin = 0, y_origin = 0),
    type = 'home_work_manhattan'))))
xsdName  <- file.path(rootPath, 
  'extdata/metadata/input_files/schema_definition', 'simulation_dict.xsd')
newSimInputFile <- file.path(Sys.getenv('R_USER'), 'newSimulation.xml')
update_SimInput(
   xmlSimInput = xmlSimInput, 
   newParam = newParam.lst, 
   xsdName  = xsdName,
   newFileName = newSimInputFile)

}
