% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSimInput.R
\name{updateSimInput}
\alias{updateSimInput}
\title{Update simulation xml input files.}
\usage{
updateSimInput(
  xmlSimInput,
  newParam,
  xsdName,
  configParamList,
  newFileName = NULL
)
}
\arguments{
\item{xmlSimInput}{Object of class \code{xml_document} from package
\code{xml2} or of class \code{character} specifying the path of the
simulation input xml file to update.}

\item{newParam}{\code{list} with the new values of the parameters to be
updated}

\item{xsdName}{}

\item{configParamList}{}

\item{newFileName}{}
}
\description{
This function takes an xml input file for the simulator and
updates it according to the set of new parameters specified in the input
parameter list \code{newParam}. The schema definition of the xml file is also
required as an input parameter to validate the structure of the output file.
}
\details{

}
\examples{
 
rootPath <- system.file(package = 'simutils')
config   <- list(Path = list(XML_VALIDATOR = file.path(rootPath, "bin")))

xmlSimInput  <- file.path(rootPath, 'extdata/input_files', 'simulation.xml')
newParam.lst <- list(
 end_time = 11,
 movement_pattern = structure(list(
    trend_angle_1_distribution = structure(
        list(mean = 140, sd = 18),
        type = 'Normal')),
    type = 'random_walk_closed_map'))
xsdName  <- file.path(rootPath, 'extdata/metadata/input_files/schema_definition', 'simulation_dict.xsd')
newSimInputFile <- file.path(Sys.getenv('R_USER'), 'newSimulation.xml')
updateSimInput(
xmlSimInput = xmlSimInput, 
newParam = newParam.lst, 
xsdName  = xsdName,
configParamList = config,
newFileName = newSimInputFile)->x

}
