% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_csv.R
\name{validate_csv}
\alias{validate_csv}
\title{Validate a csv file using an xml file as schema.}
\usage{
validate_csv(xmlFileName, csvFileName)
}
\arguments{
\item{xmlFileName}{File name of the xml file containing the schema for the
csv file.}

\item{csvFileName}{File name of the csv file to validate.}
}
\description{
\code{validate_csv} validates a csv file using an xml file as
schema, i.e. specifying its column names and value ranges (not yet
implemented).
}
\details{
Return \code{TRUE} if the csv file is validated, otherwise, it
returns \code{FALSE}.
}
\examples{
rootPath <- file.path(system.file(package = "simutils"), 'extdata')  

# antennas file
csv_fn <- file.path(rootPath, "output_files", "antennas.csv")
xml_fn <- file.path(rootPath, "metadata/output_files", "antennas_dict.xml")
validate_csv(xml_fn, csv_fn)
 
}
