% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5files.R
\name{write_h5files}
\alias{write_h5files}
\alias{.write_h5files}
\title{Write R object to .h5 Files}
\usage{
write_h5files(data, file_path = NULL)

.write_h5files(h5file, data, name = NULL)
}
\arguments{
\item{data}{The data remain to be write}

\item{file_path}{The file path of the h5file. Default is NULL and h5 file will
be generated automatically}

\item{h5file}{A new h5 file}

\item{name}{Default is NULL. One of the names of a list.}
}
\value{
An h5 file
}
\description{
This function facilitates the progress of converting any R object into .h5 file
to store information and read it in Python easily.
}
\examples{
# Generate a matrix
set.seed(1)
ref_data <- matrix(rpois(n = 10^4, lambda = 0.5), nrow = 1000)
colnames(ref_data) <- paste0("cell_", 1:ncol(ref_data))
rownames(ref_data) <- paste0("gene_", 1:nrow(ref_data))

# Generate a sparse matrix
sparse_matrix <- Matrix::Matrix(data = ref_data, sparse = TRUE)

# Generate a dataframe
df <- as.data.frame(sparse_matrix)

# Generate a list
x <- list("length" = 2,
          "integer" = 2L,
          "logical" = TRUE,
          "sparse_matrix" = sparse_matrix,
          "df" = df,
          "matrix" = ref_data)

# Set a path
file_path <- tempfile(fileext = ".h5")
# Write an h5 file
write_h5files(x, file_path = file_path)
# Load the h5 file
file.h5 <- hdf5r::H5File$new(file_path, mode = "r")
}
