% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_attrs2dt.R
\name{xmlTypes2RTypes}
\alias{xmlTypes2RTypes}
\alias{xml_attrs2dt}
\title{Create a data.table from a file metadata xml.}
\usage{
xmlTypes2RTypes(xmlType)

xml_attrs2dt(xmlname, dataset)
}
\arguments{
\item{xmlType}{character vector with the name of the type of an xml element}

\item{xmlname}{character vector of length 1 with the name of the xml file}

\item{dataset}{character vector of length 1 with the dataset name}
}
\description{
Create a data.table from a file metadata xml.
}
\examples{
xmlname <- system.file(
  "extdata/metadata/output_files", "antennas_dict.xml",
  package = "simutils"
)
xml_attrs2dt(xmlname, "antennas")

xmlname <- system.file(
  "extdata/metadata/output_files", "persons_dict.xml",
  package = "simutils"
)
xml_attrs2dt(xmlname, "individuals")

}
