% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_getters.R
\name{getCoordsNames}
\alias{getCoordsNames}
\alias{getGridNoTilesX}
\alias{getGridNoTilesY}
\alias{getSignalType}
\alias{getXTileDimColName}
\alias{getYTileDimColName}
\alias{getCellCoordName}
\alias{getCellIDName}
\alias{getSpatialUnitName}
\alias{getNestingSpatialUnitName}
\title{Get different elements from an xml file.}
\usage{
getCoordsNames(xmlname, dataset)

getGridNoTilesX(xmlname, dataset)

getGridNoTilesY(xmlname, dataset)

getSignalType(xmlname, dataset)

getXTileDimColName(xmlname, dataset)

getYTileDimColName(xmlname, dataset)

getCellCoordName(xmlname, dataset)

getCellIDName(xmlname, dataset)

getSpatialUnitName(xmlname, dataset)

getNestingSpatialUnitName(xmlname, dataset)
}
\arguments{
\item{xmlname}{character vector of length 1 with the name of the xml file}

\item{dataset}{character vector of length 1 with the dataset name}
}
\description{
Get different elements from an xml file.
}
\examples{
xmlname <- system.file(
  "extdata/metadata/output_files/antennas_dict.xml", package = "simutils") 
getCoordsNames(xmlname, 'antennas')

xmlname <- system.file(
  "extdata/metadata/output_files/persons_dict.xml", package = "simutils") 
getCoordsNames(xmlname, 'individuals')

xmlname <- system.file(
  "extdata/metadata/output_files/grid_dict.xml", package = "simutils") 
getGridNoTilesX(xmlname, 'grid')
getGridNoTilesY(xmlname, 'grid')

}
