% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_effect}
\alias{add_effect}
\title{Shift Specific Target Values}
\usage{
add_effect(data, params, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{params}{named vector that maps parameter names to parameter values}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\description{
In order to simulate power of a statistical test, a random effect is added
to the (permuted) target variable values. Thus, a situation in which H1
holds is established. The distribution of the additive effect is specified
by user command line arguments together with a config object. Depending on
the (user-specified) target variable, the respective values will be
truncated after the effects have been added.
}
\details{
\code{data} is modified in place.

\code{options$scenario} determines the simulation scenario. If
\code{options$scenario == 1}, only independent main effects are added.
\code{options$scenario == 2} means that, in addition to the main effects, also
dependent effects are added.

\code{options$target} contains the name of the target variable in \code{data}.
\code{config$max_values} contains a named vector with maximum values for various
target variables.
Moreover, \code{options} and \code{config} must contain all entries required by
\code{add_main_effect} and \code{add_dependent_effect}.
}
