% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_main_effect}
\alias{add_main_effect}
\title{Add Independent Random Effect to Target}
\usage{
add_main_effect(data, params, options, config)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{params}{named vector that maps parameter names to parameter values}

\item{options}{list with user-defined command line arguments}

\item{config}{list with further arguments}
}
\value{
vector with the effects that were added to the target variable
}
\description{
At a specific group and at specific points of time, add random integer
effects to the target variable.
}
\details{
options$effect_type is the name of the distribution of the added effects.
The params argument provides additional parameters for this distribution.

options$target is the user-defined name of the target variable in data to
which effects are applied.

config$main_effect_time is a vector that specifies the points of time at
which effects are added.
config$time_variable is the name of the column in data that contains times.

config$placebo_group is the name of the group to which effects are added.
config$group_variable is the name of the column in data that contains groups.

Note that data will be modified in place.
}
