% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{binarize_target}
\alias{binarize_target}
\title{Dichotomize Target Variable}
\usage{
binarize_target(binarize, data, target, blocklength, threshold)
}
\arguments{
\item{binarize}{whether to apply binarization or not (TRUE/FALSE)}

\item{data}{data.table with the simulation data}

\item{target}{name of the target column in data}

\item{blocklength}{number of measurements in a block}

\item{threshold}{decides whether a value is replaced by 0 or 1}
}
\description{
If desired, replace the target variable values with 0/1 values.
}
\details{
Target variable values are supposed to be divided into blocks of equal size.
Each block combines measurements for one subject at different points of
time. Whether a measurement is replaced by 0 or by 1 depends on its
magnitude relative to the first value in the same block.

Note that data is modified in place.
}
