% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.r
\name{boot}
\alias{boot}
\title{Create a Bootstrap Data Object}
\usage{
boot(
  fun,
  train,
  test,
  B,
  type_boot = c("bayesian", "nonparametric"),
  fun_out_boot
)
}
\arguments{
\item{fun}{A function that fits a model to training data and returns coefficients.}

\item{train}{The training dataset.}

\item{test}{The test dataset.}

\item{B}{The number of bootstrap samples to generate.}

\item{type_boot}{The type of bootstrap to perform ("bayesian" or "nonparametric").}

\item{fun_out_boot}{A vector of names indicating which elements of the function output to bootstrap.}
}
\value{
A bootstrap object with the bootstrapped coefficient estimates.
}
\description{
This function creates a bootstrap data object from a model-fitting function, using either nonparametric or Bayesian bootstrap methods.
}
