% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_batch_properties.R
\name{calculate_batch_properties}
\alias{calculate_batch_properties}
\title{Calculate Batch Quantifying Metrics}
\usage{
calculate_batch_properties(
  data,
  batch_info,
  cluster_info = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{A matrix of gene expression profile.}

\item{batch_info}{Batch assignment of every cells in columns of matrix.}

\item{cluster_info}{Cluster(or group) assignment of every cells in columns of matrix.}

\item{verbose}{Whether messages are returned during the process.}
}
\description{
Calculate Batch Quantifying Metrics
}
