% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_parameters.R
\name{change_values_in_list}
\alias{change_values_in_list}
\title{Change Values From a List}
\usage{
change_values_in_list(list, new_values)
}
\arguments{
\item{list}{The list.}

\item{new_values}{Other attr ibutes and new values.}
}
\description{
Change Values From a List
}
