% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_alpha_error}
\alias{compute_alpha_error}
\title{Simulation-based Computation of the Type-I Error}
\usage{
compute_alpha_error(data, options, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{options}{\code{list} with user-defined command line arguments}

\item{config}{\code{list} with further arguments}
}
\value{
vector with average type-I errors for both periods
}
\description{
For a given number of repetitions, permute the target variable to establish
a situation in which H0 holds. Perform hypotheses tests for both periods of
the trial separately. Store all test results and return the average for both
periods.
}
\details{
\code{options$target} contains the name of the target variable.
\code{config$repetitions} is the number of repetitions to perform (i.e., the
number of tests performed for each period)
\code{config$blocklength} is the number of measurements in a block that refer to
one subject.
\code{config$alpha} is the expected type I error rate.

Moreover, \code{options} and \code{config} must contain all attributes required by
\code{test_h0}.
}
