% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simElements.R
\name{create_simElements}
\alias{create_simElements}
\title{Create list of parameters of the simulation.}
\usage{
create_simElements(filenames, crs = NA_integer_)
}
\arguments{
\item{filenames}{Named list of filenames with each element of the simulation.}

\item{crs}{integer or character; coordinate reference system for the geometry
as in function \code{sf::st_as_sfc}}
}
\description{
Create a named list of different elements for the simulation. It basically
contains the map (regions included), the network configuration (position of
BTS, power, attenuation factor, etc.), the grid, the radio cells associated
to each antenna, and the telecommunication signal measures (RSS, SDM, ...).
}
\examples{
filename_map      <- c(
 xml= system.file("extdata/input_files", "map.xml", package = "simutils"),
 xsd= '')
  
filename_network  <- c(
 csv= system.file("extdata/output_files/antennas.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/antennas_dict.xml", 
                   package = "simutils"))
                   
filename_signal <- c(
 csv= system.file("extdata/output_files/SignalMeasure_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/SignalMeasure_dict.xml", 
                   package = "simutils"))
                 
filename_coverage <- c(
 csv= system.file("extdata/output_files", "AntennaCells_MNO1.csv", 
                  package = "simutils"),
 xml= system.file("extdata/metadata/output_files/AntennaCells_dict.xml", 
                  package = "simutils"))
                       
filename_grid <- c(
  csv= system.file("extdata/output_files/grid.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/grid_dict.xml", 
                   package = "simutils")) 

filename_individ <- c(
  csv= system.file("extdata/output_files/persons.csv", package = "simutils"),
  xml= system.file("extdata/metadata/output_files/persons_dict.xml", 
                   package = "simutils"))   
                       
filenames <- list(
  map                = filename_map,
  network_parameters = filename_network,
  signal             = filename_signal,
  coverage_cells     = filename_coverage,
  grid               = filename_grid,
  individuals        = filename_individ)
  
simElements <- create_simElements(filenames, crs = 2062)
str(simElements)

}
