% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_period_info}
\alias{get_period_info}
\title{Summarise Dataset Containing a Single Trial Period}
\usage{
get_period_info(data, config)
}
\arguments{
\item{data}{\code{data.table} with the study data}

\item{config}{global config object

\code{config$subject_variable} is the name of the variable that identifies
subjects in data.
\code{config$group_variable} is the name of the group variable in \code{data}.}
}
\value{
printable string with dataset info.
}
\description{
Create a (multiline) string with information about the given dataset.
}
