% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_S4_values.R
\name{get_S4_values}
\alias{get_S4_values}
\title{Get Values From S4 Object}
\usage{
get_S4_values(S4Object)
}
\arguments{
\item{S4Object}{A S4object}
}
\value{
A list
}
\description{
Get Values From S4 Object
}
