% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_split_info}
\alias{get_split_info}
\title{Summarize Dataset Split According to Trial Period}
\usage{
get_split_info(data, config)
}
\arguments{
\item{data}{data.table with the study data}

\item{config}{global config object

config$first_period_end is the last timepoint in the first trial period.
config$time_variable is the name of the variable containing timepoints in
the dataset.
config$subject_variable is the name of the variable that identifies subjects
in data.
config$group_variable is the name of the group variable in data.}
}
\value{
printable string with dataset info.
}
\description{
Create a (multiline) string with information about the given dataset.
}
