% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpc.R
\name{gpc}
\alias{gpc}
\title{Perform Hypothesis Test using Generalized Pairwise Comparisons (GPC)}
\usage{
gpc(data, type, repeated, matching, best, options, config, verbose = FALSE)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{type}{of the GPC ("univariate", (multivariate) "prioritized",
(multivariate) "non-prioritized")}

\item{repeated}{vector of (prioritized order of) repeated measures (provided
in a column named Time in dataset)}

\item{matching}{"matched" or "unmatched" GPC}

\item{best}{"higher" ("lower") if higher (lower) values are the preferred
outcome}

\item{options}{\code{list} with user-defined command line arguments (among others
the target)}

\item{config}{\code{list} with further arguments}

\item{verbose}{\code{logical} value indication whether wins/losses/ties and net-
benefit should be printed.}
}
\value{
\code{list} of p-values for the respective tests
}
\description{
Perform Hypothesis Test using Generalized Pairwise Comparisons (GPC)
}
