% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_similarity.R
\name{leaf_node_depth}
\alias{leaf_node_depth}
\title{Trace the Depths of All Leaf Nodes}
\usage{
leaf_node_depth(edge)
}
\arguments{
\item{edge}{The edge element in Phylo tree object}
}
\value{
A numeric vector of every leaf node
}
\description{
Trace the Depths of All Leaf Nodes
}
