% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deepest.R
\name{list_deepest}
\alias{list_deepest}
\title{Transform an object and its elements in a list of lists keeping
the attributes' structure.}
\usage{
list_deepest(object)
}
\arguments{
\item{object}{A \code{list} or another object with an attributes' structure.}
}
\description{
\code{list_deepest} receives an object and transforms it
and all its elements in a list of lists keeping the original attributes'
structure.
}
\details{
Return a \code{list} of \code{list}s using the attributes' structure
from the original object.
}
\examples{
rootPath  <- system.file(package = "simutils")  
xmlFile   <- file.path(rootPath, "extdata/input_files", "simulation.xml")
xmlObject <- xml2::as_list(xml2::read_xml(xmlFile))
list_deepest(xmlObject)

}
