% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{meta_info}
\alias{meta_info}
\title{Define Meta Information of A Dataset}
\usage{
meta_info(
  id,
  repository = NULL,
  accession_number = NULL,
  URL = NULL,
  platform,
  species,
  organ = NULL,
  cell_num,
  gene_num,
  data_type = "count",
  ERCC = FALSE,
  dilution_factor = NULL,
  volume = NULL,
  group_condition = NULL,
  treatment = NULL,
  batch_info = NULL,
  cluster_info = NULL,
  spatial_coordinate = NULL,
  start_cell = NULL
)
}
\arguments{
\item{id}{Dataset ID}

\item{repository}{Which repository or cloud contains the dataset, e.g. GEO, ArrayExpress.}

\item{accession_number}{The accession number in the repository if available.}

\item{URL}{The URL to downloiad the dataset.}

\item{platform}{Sequencing platform.}

\item{species}{Organism.}

\item{organ}{Which source or organ do the samples come from?}

\item{cell_num}{The cell number of the dataset.}

\item{gene_num}{The gene number of the dataset.}

\item{data_type}{Which type is the dataset? e.g. count, FPKM.}

\item{ERCC}{Whether the dataset contains ERCC spike-in genes?}

\item{dilution_factor}{If there are spike-in genes, dilution factor is neccessary.}

\item{volume}{The volume (nanoliter) of mix liquid used in experiment.}

\item{group_condition}{Vector. The group assignment of cells in the dataset.}

\item{treatment}{Vector. The treatment of cells in the experiment.}

\item{batch_info}{Vector. The batch information of cells in the experiment.}

\item{cluster_info}{Vector. The cluster information of cells.}

\item{spatial_coordinate}{A data frame of x and y coordinates of spatial transcriptomics data}

\item{start_cell}{The cell id which is determined as the start_cell of a trajectory}
}
\value{
A list of dataset information.
}
\description{
Define Meta Information of A Dataset
}
