% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_Boolean}
\alias{param_Boolean}
\title{Define Logical Parameters}
\usage{
param_Boolean(
  id,
  type = "logical",
  default,
  force = FALSE,
  description = NULL,
  process = "simulation"
)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be a logical value.}

\item{default}{The default value. TRUE or FALSE.}

\item{force}{Logical, whether this parameter must be needed or not.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}
}
\value{
A list.
}
\description{
Define a Boolean parameter.
}
\examples{
verbose <- param_Boolean(
id = "verbose",
default = TRUE,
force = TRUE,
description = "Whether to return the information during the process"
)
}
