% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_character}
\alias{param_character}
\title{Define Character Parameters}
\usage{
param_character(
  id,
  type = "character",
  default,
  alternatives,
  force = FALSE,
  description = NULL,
  process = "simulation",
  function_name
)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be characters.}

\item{default}{The default values. Length to one or more.}

\item{alternatives}{Options that can be chosen.}

\item{force}{Logical, whether this parameter must be needed or not.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}

\item{function_name}{The function name where the parameter exists.}
}
\value{
A list.
}
\description{
Define a character parameter.
}
\examples{
dimension_method <- param_character(
id = 'dimension_method',
default = "UMAP",
process = "estimation",
alternatives = c("UMAP", "TSNE", "PCA"),
force = TRUE,
description = "The algorithm to use for dimension reduction."
)
}
