% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_character}
\alias{param_character}
\title{Define Character Parameters}
\usage{
param_character(
  id,
  type = "character",
  default,
  alternatives,
  description,
  process = "simulation"
)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be characters.}

\item{default}{The default values. Length to one or more.}

\item{alternatives}{Options that can be chosen.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}
}
\value{
A list.
}
\description{
Define a character parameter.
}
\examples{
dimension_method <- param_character(id = 'dimension_method',
                                    default = "UMAP",
                                    process = "estimation",
                                    alternatives = c("UMAP", "TSNE", "PCA"),
                                    description = "The algorithm to use for
                                                   dimension reduction.")
}
