% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_numeric}
\alias{param_numeric}
\title{Define Numeric Parameters}
\usage{
param_numeric(
  id,
  type = "numeric",
  default,
  lower = NULL,
  upper = NULL,
  border = TRUE,
  description,
  process = "simulation"
)
}
\arguments{
\item{id}{A parameter name.}

\item{type}{The type of parameter. Must be a numeric.}

\item{default}{The default value.}

\item{lower}{The lowest range of this parameter.}

\item{upper}{The highest range of this parameter.}

\item{border}{Boolean. Whether the default value can get to the range border.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}
}
\value{
A list.
}
\description{
Define a numeric parameter.
}
\examples{
probability <- param_numeric(id = "probability",
                             default = 0.5,
                             lower = 0,
                             upper = 1,
                             border = TRUE,
                             description = "The probability to select the right number")
}
