% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{param_others}
\alias{param_others}
\title{Define Other Object Parameters}
\usage{
param_others(
  id = NULL,
  type = NULL,
  default = NULL,
  description = NULL,
  process = "simulation"
)
}
\arguments{
\item{id}{A parameter name. Default is NULL.}

\item{type}{The type of parameter, which can be a list or any other object you input.}

\item{default}{The default input.}

\item{description}{The description information of this parameter.}

\item{process}{Two options, simulation or estimation. Which process does this
parameter involved in.}
}
\value{
A list.
}
\description{
Define Other Object Parameters
}
\examples{
list_input <- param_others(
id = 'input_list',
type = "list",
process = "simulation"
)
}
