% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_data}
\alias{read_data}
\title{Read in Diacerin Study Dataset}
\usage{
read_data(filename, config)
}
\arguments{
\item{filename}{path to the tab-separated dataset file}

\item{config}{list with further arguments}
}
\value{
the preprocessed dataset as data.table
}
\description{
The dataset is read in as data.table. This is important as many of the other
functions require this data type. The dataset is supposed to consist of
tab-separated values (tsv). Its rows will be reordered according to
config$subject_variable and config$time_variable. Moreover, the time
variable will be casted to numeric to establish an order w.r.t. time.
}
