% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_map.R
\name{read_xml_map}
\alias{read_xml_map}
\title{Read string in WKT format and convert it into an sf object.}
\usage{
read_xml_map(xmlname)
}
\arguments{
\item{text}{Character string of WKT}

\item{id}{Character vector of unique ids to label geometries as in function
\code{rgeos::readWKT}. Length must match the number of subgeometries in the
WKT}

\item{p4s}{Either a character string or an object of class \code{CRS} as in
the function \code{rgeos::readWKT}}

\item{...}{passed on to \code{sf::st_as_sf}, might included named arguments
crs or precision}
}
\description{
Read the input WKT string, parse it, and convert it into an sf object.
}
\examples{
read_xml_map(system.file(
   "extdata/input_files", "map.xml", package = "simutils"
))

}
