% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.r
\name{robust}
\alias{robust}
\title{Create a Robust Data Object}
\usage{
robust(model, regression_name, suppressWarnings = F)
}
\arguments{
\item{model}{A linear model object}

\item{regression_name}{A name for the regression model}

\item{suppressWarnings}{A logical value to suppress warnings}
}
\value{
A robust object
}
\description{
This function creates a robust data object from a linear model
}
