% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDockerImage.R
\name{runDockerImage}
\alias{runDockerImage}
\title{Run the simulation software using the docker image.}
\usage{
runDockerImage(
  dockerImage,
  input_folder,
  simulationCFGFile,
  mapFile,
  personsCFGFile,
  antennasCFGFile,
  output_folder
)
}
\arguments{
\item{dockerImage}{Absolute path where the file to run the docker image is.}

\item{input_folder}{Absolute path of the folder which contains the input data.}

\item{simulationCFGFile}{Name of the simulation configuration file.}

\item{mapFile}{Name of the map file.}

\item{personsCFGFile}{Name of the persons' configuration file.}

\item{antennasCFGFile}{Name of the antennas' configuration file.}

\item{output_folder}{Name of the output folder indicated in the
simulation's configuration file.}
}
\description{
\code{runDockerImage} runs the simulation software using
the docker image.
}
\details{
Return invisible NULL after executing the docker image.
}
\examples{
rootPath          <- system.file(package = 'simutils')
dockerImage       <- file.path(rootPath, "extdata/docker/simulator.tar")
input_folder      <- file.path(rootPath, "extdata/input_files")
simulationCFGFile <- 'simulation.xml'
mapFile           <- 'map.wkt'
personsCFGFile    <- 'persons.xml'
antennasCFGFile   <- 'antennas.xml'
output_folder     <- file.path(Sys.getenv('R_USER'), 'example_docker')
runDockerImage(
  dockerImage = dockerImage,
  input_folder = input_folder,
  simulationCFGFile = simulationCFGFile,
  mapFile = mapFile,
  personsCFGFile = personsCFGFile,
  antennasCFGFile = antennasCFGFile,
  output_folder = output_folder
)

}
