% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExeFile.R
\name{runExeFile}
\alias{runExeFile}
\title{Run an MNO simulation using the simulator.exe file.}
\usage{
runExeFile(
  configParamList,
  sysinfo,
  input_folder,
  simulationCFGFile,
  mapFile,
  personsCFGFile,
  antennasCFGFile,
  outputFolder
)
}
\arguments{
\item{configParamList}{\code{list} which contains the different paths needed
during the simulation. It must be read from the config.ini file.}

\item{sysinfo}{\code{character} with the info about the system where
the simulation is going to be executed such as the OS, the username...}

\item{input_folder}{Absolute path where the input files are.}

\item{simulationCFGFile}{Name of the configuration simulation file
(input file).}

\item{mapFile}{Name of the map file (input file).}

\item{personsCFGFile}{Name of the persons file (input file).}

\item{antennasCFGFile}{Name of the antennas file (input file).}

\item{outputFolder}{Full path where the output files are going to be saved.}
}
\description{
\code{runExeFile} reads the simulator path to exe from the
config.ini file and execute the simulator.exe from this path.
If the file doesn't exist, it returns an error.
}
\details{
Return invisible NULL after placing output files of the simulation
in the output folder.
}
\examples{
config        <- list(
  Path = list(
    SIMULATOR_PATH_TO_EXE = file.path(system.file(package = "simutils"), "inst/bin")))
sysinfo           <- Sys.info()
input_folder      <- file.path(system.file(package = "simutils", "inst/extdata/input_files"))
simulationCFGFile <- "simulation.xml"
mapFile           <- "map.wkt"
personsCFGFile    <- "persons.xml"
antennasCFGFile   <- "antennas.xml"
outputFolder      <- Sys.getenv('R_USER')
runExeFile(
  configParamList   = config,
  sysinfo           = sysinfo,
  input_folder      = input_folder,
  simulationCFGFile = simulationCFGFile,
  mapFile           = mapFile,
  personsCFGFile    = personsCFGFile,
  antennasCFGFile   = antennasCFGFile,
  outputFolder      = antennasCFGFile)
}
