% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExeFile.R
\name{runExeFile}
\alias{runExeFile}
\title{Run an MND simulation using the simulator.exe file.}
\usage{
runExeFile(
  path_to_exe,
  simulator_version,
  input_folder,
  simulationCFGFile,
  mapFile,
  personsCFGFile,
  antennasCFGFile
)
}
\arguments{
\item{path_to_exe}{character vector with the absolute path to the executable file of the
simulator of mobile network event data.}

\item{simulator_version}{character denoting the version of the simulator (e.g. "1.2.0")}

\item{input_folder}{absolute path where the simulation input files are.}

\item{simulationCFGFile}{name of the configuration simulation file (input file).}

\item{mapFile}{name of the map file (input file).}

\item{personsCFGFile}{name of the persons file (input file).}

\item{antennasCFGFile}{name of the antennas file (input file).}
}
\description{
\code{runExeFile} reads the simulator path to the executable file of the simulator
of mobile network event data and the input file parameters to execute the simulator.exe from this
path.

If the executable file doesn't exist, the function will donwload it automatically from the
Assests  section of this URL in github.com (for version 1.2.0):

https://github.com/bogdanoancea/simulator/releases/tag/1.2.0

The name of the executable file is \code{simulator.exe}.

For Windows, the user can download an installer from the URL (for version 1.2.0):

https://github.com/bogdanoancea/simulator/releases/tag/1.2.0-kit
}
\details{
Return invisible \code{NULL} after placing output files of the
simulation in the output folder.
}
\examples{
\dontrun{
# simulator.exe will be downloaded from our github repo
runExeFile(
 path_to_exe       = "",  
 simulator_version = "1.2.0",
 input_folder      = file.path(rootPath, "extdata/input_files"),
 simulationCFGFile = "simulation.xml",
 mapFile           = "map.wkt",
 personsCFGFile    = "persons.xml",
 antennasCFGFile   = "antennas.xml")
 }
  
}
