% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarize_tests}
\alias{summarize_tests}
\title{Summarize Hypothesis Tests}
\usage{
summarize_tests(results_df)
}
\arguments{
\item{results_df}{\code{data.frame} with columns of test results (logical vectors)}
}
\value{
summary list
}
\description{
compute H0 rejection rate and count number of failed tests based on a
data.frame of test results that contains \code{TRUE} for each rejected H0,
\code{FALSE} for each non-rejected H0 and \code{NA} for each failed test.
}
