% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{truncate_target}
\alias{truncate_target}
\title{Truncate Target Variable}
\usage{
truncate_target(data, target, max_values)
}
\arguments{
\item{data}{data.table with the simulation data}

\item{target}{name of the target column in data}

\item{max_values}{named vector that maps variable names to maximum values}
}
\description{
To ensure that the target variable values do not exceed their scale they can
be truncated with this function.
}
\details{
Truncation is only applied to the variables specified by the names of
max_values.

Note that data is modified in place.
}
