% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{truncate_target}
\alias{truncate_target}
\title{Truncate Target Variable}
\usage{
truncate_target(data, target, config)
}
\arguments{
\item{data}{\code{data.table} with the simulation data}

\item{target}{name of the target variable}

\item{config}{\code{list} with further arguments}
}
\description{
To ensure that the target variable values do not exceed their scale they are
truncated in this function.
}
\details{
Truncation is only applied to the variables specified by the names of
\code{config$max_values} or \code{config$min_values}.

Note that \code{data} is modified in place.
}
