% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSimInput.R
\name{updateSimInput}
\alias{updateSimInput}
\title{Update simulation xml input files.}
\usage{
updateSimInput(xmlSimInput, newParam, xsdName, configParamList, newFile = NULL)
}
\arguments{
\item{xmlSimInput}{}

\item{newParam}{}

\item{xsdName}{}

\item{configParamList}{}

\item{newFile}{}
}
\description{

}
\details{

}
\examples{
config        <- list(
  Path = list(
    SIMULATOR_PATH_TO_EXE = file.path(system.file(package = "simutils"), "bin")))
sysinfo           <- Sys.info()
input_folder      <- file.path(system.file(package = "simutils", "extdata/input_files"))
simulationCFGFile <- "simulation.xml"
mapFile           <- "map.wkt"
personsCFGFile    <- "persons.xml"
antennasCFGFile   <- "antennas.xml"
outputFolder      <- Sys.getenv('R_USER')
runExeFile(
  configParamList   = config,
  sysinfo           = sysinfo,
  input_folder      = input_folder,
  simulationCFGFile = simulationCFGFile,
  mapFile           = mapFile,
  personsCFGFile    = personsCFGFile,
  antennasCFGFile   = antennasCFGFile,
  outputFolder      = outputFolder)
 


rootPath <- system.file(package = 'simutils')
xmlSimInput  <- file.path(rootPath, 'extdata/input_files', 'simulation.xml')
newParam.lst <- list(
 end_time = 11,
 movement_pattern = structure(list(
    trend_angle_1_distribution = structure(
        list(mean = 140, sd = 18),
        type = 'Normal')),
    type = 'random_walk_closed_map'))
xsdName  <- file.path(rootPath, 'extdata/metadata/input_files/schema_definition', 'simulation_dict.xsd')
newSimInputFile <- file.path(Sys.getenv('R_USER'), 'newSimulation.xml')
updateSimInput(
xmlSimInput = xmlSimInput, 
newParam = newParam.lst, 
xsdName  = xsdName,
newFile = newSimInputFile)
}
